Extract service-call dispatch into NodeManagerServiceDispatcher
Move the session-service dispatch implementations (GetManagerHandle,
TranslateBrowsePaths, Browse/BrowseNext, Read, Write, Call, history
access, ConditionRefresh, monitored-item create/restore/modify/
transfer/delete/set-mode, and request validation) out of
MasterNodeManager into the internal NodeManagerServiceDispatcher,
which depends only on the read-only routing snapshot, the server
context, and the notification-dispatch lease accessors. Lifecycle
mutation state stays private to MasterNodeManager and is compiler-
inaccessible from the dispatcher.

MasterNodeManager keeps the entire public surface unchanged in the
new MasterNodeManager.ServiceDispatch.cs partial: every public
virtual service method, protected helper, obsolete wrapper, and
static validator retains its signature, virtual-ness, and docs, and
delegates to the dispatcher. Internal dispatch calls back through
the owner's virtual GetManagerHandleAsync so subclass overrides
keep affecting handle resolution.

Follows up #4094; part of #4098.
