+++ src/Opc.Ua.Server/NodeManager/NodeManagerServiceDispatcher.cs
@@ -1,0 +1,3522 @@
+/* ========================================================================
+ * Copyright (c) 2005-2025 The OPC Foundation, Inc. All rights reserved.
+ *
+ * OPC Foundation MIT License 1.00
+ *
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ *
+ * The complete license agreement can be found here:
+ * http://opcfoundation.org/License/MIT/1.00/
+ * ======================================================================*/
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using NotificationDispatchLease = Opc.Ua.Server.MasterNodeManager.NotificationDispatchLease;
+using RetiredGenerationNotifications = Opc.Ua.Server.MasterNodeManager.RetiredGenerationNotifications;
+
+namespace Opc.Ua.Server
+{
+    /// <summary>
+    /// Routes OPC UA session service calls (Browse, TranslateBrowsePaths, Read, Write,
+    /// Call, history access, and monitored item operations) to the NodeManagers that own
+    /// the requested Nodes, together with the per request validation that dispatch needs.
+    /// <para>
+    /// Boundary contract: this class depends on the read-only routing snapshot
+    /// (<see cref="NodeManagerRoutingTable"/> reads), the server context, and the
+    /// notification-dispatch lease accessors that <see cref="MasterNodeManager"/>
+    /// exposes for retired generations. It must not acquire lifecycle semaphores,
+    /// mutate routing state, or use <see cref="IDynamicNodeManagerHost"/> - lifecycle
+    /// coordination stays with <see cref="MasterNodeManager"/>, whose private mutation
+    /// state is inaccessible from here by construction. Calls back into
+    /// <see cref="MasterNodeManager"/> public virtual methods (handle resolution) keep
+    /// subclass overrides effective for internal dispatch.
+    /// </para>
+    /// </summary>
+    internal sealed class NodeManagerServiceDispatcher
+    {
+        /// <summary>
+        /// Initializes the dispatcher over the owning master node manager's routing table.
+        /// </summary>
+        public NodeManagerServiceDispatcher(
+            MasterNodeManager owner,
+            NodeManagerRoutingTable nodeManagers,
+            IServerInternal server,
+            ILogger logger)
+        {
+            m_owner = owner ?? throw new ArgumentNullException(nameof(owner));
+            m_nodeManagers = nodeManagers ?? throw new ArgumentNullException(nameof(nodeManagers));
+            Server = server ?? throw new ArgumentNullException(nameof(server));
+            m_logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        }
+
+        /// <summary>
+        /// The server the owning master node manager belongs to.
+        /// </summary>
+        private IServerInternal Server { get; }
+
+        /// <summary>
+        /// The visible NodeManagers serving each namespace index.
+        /// </summary>
+        private IReadOnlyDictionary<int, IReadOnlyList<IAsyncNodeManager>> NamespaceManagers
+            => m_nodeManagers.NamespaceManagers;
+
+        internal object? GetManagerHandle(NodeId nodeId, out INodeManager? nodeManager)
+        {
+            object? handle;
+            nodeManager = null;
+
+            // null node ids have no manager.
+            if (nodeId.IsNull)
+            {
+                return null;
+            }
+
+            // use the namespace index to select the node manager.
+            int index = nodeId.NamespaceIndex;
+
+            // check if node managers are registered - use the core node manager if unknown.
+            if (!NamespaceManagers.TryGetValue(index, out IReadOnlyList<IAsyncNodeManager>? nodeManagers))
+            {
+                handle = m_nodeManagers[1].SyncNodeManager.GetManagerHandle(nodeId);
+
+                if (handle != null)
+                {
+                    nodeManager = m_nodeManagers[1].SyncNodeManager;
+                    return handle;
+                }
+                return null;
+            }
+
+            foreach (IAsyncNodeManager asyncNodeManager in nodeManagers!)
+            {
+                handle = asyncNodeManager.SyncNodeManager.GetManagerHandle(nodeId);
+
+                if (handle != null)
+                {
+                    nodeManager = asyncNodeManager.SyncNodeManager;
+                    return handle;
+                }
+            }
+
+            // node not recognized.
+            return null;
+        }
+
+        internal async ValueTask<(object? handle, IAsyncNodeManager? nodeManager)>
+            GetManagerHandleAsync(NodeId nodeId, CancellationToken cancellationToken = default)
+        {
+            object? handle;
+
+            // null node ids have no manager.
+            if (nodeId.IsNull)
+            {
+                return (null, null);
+            }
+
+            // use the namespace index to select the node manager.
+            int index = nodeId.NamespaceIndex;
+
+            // check if node managers are registered - use the core node manager if unknown.
+            if (!NamespaceManagers.TryGetValue(index, out IReadOnlyList<IAsyncNodeManager>? nodeManagers))
+            {
+                handle = await m_nodeManagers[1].GetManagerHandleAsync(nodeId, cancellationToken)
+                    .ConfigureAwait(false);
+
+                if (handle != null)
+                {
+                    return (handle, m_nodeManagers[1]);
+                }
+                return (null, null);
+            }
+
+            foreach (IAsyncNodeManager asyncNodeManager in nodeManagers!)
+            {
+                handle = await asyncNodeManager.GetManagerHandleAsync(nodeId, cancellationToken)
+                    .ConfigureAwait(false);
+
+                if (handle != null)
+                {
+                    return (handle, asyncNodeManager);
+                }
+            }
+
+            // node not recognized.
+            return (null, null);
+        }
+
+        internal async ValueTask<(ArrayOf<BrowsePathResult> results, ArrayOf<DiagnosticInfo> diagnosticInfos)>
+            TranslateBrowsePathsToNodeIdsAsync(
+            OperationContext context,
+            ArrayOf<BrowsePath> browsePaths,
+            CancellationToken cancellationToken = default)
+        {
+            bool diagnosticsExist = false;
+            var results = new List<BrowsePathResult>(browsePaths.Count);
+            var diagnosticInfos = new List<DiagnosticInfo>(browsePaths.Count);
+
+            for (int ii = 0; ii < browsePaths.Count; ii++)
+            {
+                // check if request has timed out or been cancelled.
+                if (StatusCode.IsBad(context.OperationStatus))
+                {
+                    throw new ServiceResultException(context.OperationStatus);
+                }
+
+                BrowsePath browsePath = browsePaths[ii];
+
+                var result = new BrowsePathResult { StatusCode = StatusCodes.Good };
+                results.Add(result);
+
+                ServiceResult error;
+
+                // need to trap unexpected exceptions to handle bugs in the node managers.
+                try
+                {
+                    error = await TranslateBrowsePathAsync(context, browsePath, result, cancellationToken)
+                        .ConfigureAwait(false);
+                }
+                catch (Exception e)
+                {
+                    error = ServiceResult.Create(
+                        e,
+                        StatusCodes.BadUnexpectedError,
+                        "Unexpected error translating browse path.");
+                }
+
+                if (ServiceResult.IsGood(error))
+                {
+                    // check for no match.
+                    if (result.Targets.Count == 0)
+                    {
+                        error = StatusCodes.BadNoMatch;
+                    }
+                    // put a placeholder for diagnostics.
+                    else if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        diagnosticInfos.Add(null!);
+                    }
+                }
+
+                // check for error.
+                if (error != null && error.Code != StatusCodes.Good)
+                {
+                    result.StatusCode = error.StatusCode;
+
+                    if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        DiagnosticInfo? diagnosticInfo = ServerUtils.CreateDiagnosticInfo(
+                            Server,
+                            context,
+                            error,
+                            m_logger);
+                        diagnosticInfos.Add(diagnosticInfo!);
+                        diagnosticsExist = true;
+                    }
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (results, diagnosticInfos);
+        }
+
+        internal void UpdateDiagnostics(
+            OperationContext context,
+            bool diagnosticsExist,
+            ref List<DiagnosticInfo> diagnosticInfos)
+        {
+            if (diagnosticsExist && context.StringTable.Count == 0)
+            {
+                diagnosticsExist = false;
+
+                for (int ii = 0; !diagnosticsExist && ii < diagnosticInfos.Count; ii++)
+                {
+                    DiagnosticInfo? diagnosticInfo = diagnosticInfos[ii];
+
+                    int depth = 0;
+                    while (diagnosticInfo != null && depth++ < DiagnosticInfo.MaxInnerDepth)
+                    {
+                        if (!string.IsNullOrEmpty(diagnosticInfo.AdditionalInfo))
+                        {
+                            diagnosticsExist = true;
+                            break;
+                        }
+
+                        diagnosticInfo = diagnosticInfo.InnerDiagnosticInfo;
+                    }
+                }
+            }
+
+            if (!diagnosticsExist)
+            {
+                diagnosticInfos = null!;
+            }
+        }
+
+        internal async ValueTask<ServiceResult> TranslateBrowsePathAsync(
+            OperationContext context,
+            BrowsePath browsePath,
+            BrowsePathResult result,
+            CancellationToken cancellationToken)
+        {
+            Debug.Assert(browsePath != null);
+            Debug.Assert(result != null);
+
+            // check for valid start node.
+            (object? sourceHandle, IAsyncNodeManager? nodeManager) = await m_owner.GetManagerHandleAsync(
+                browsePath!.StartingNode,
+                cancellationToken)
+                .ConfigureAwait(false);
+
+            if (sourceHandle == null)
+            {
+                return StatusCodes.BadNodeIdUnknown;
+            }
+
+            // check the relative path.
+            RelativePath relativePath = browsePath.RelativePath;
+
+            if (relativePath.Elements.IsEmpty)
+            {
+                return StatusCodes.BadNothingToDo;
+            }
+
+            for (int ii = 0; ii < relativePath.Elements.Count; ii++)
+            {
+                RelativePathElement element = relativePath.Elements[ii];
+
+                if (element == null || relativePath.Elements[ii].TargetName.IsNull)
+                {
+                    return StatusCodes.BadBrowseNameInvalid;
+                }
+
+                if (element.ReferenceTypeId.IsNull)
+                {
+                    element.ReferenceTypeId = ReferenceTypeIds.References;
+                    element.IncludeSubtypes = true;
+                }
+            }
+
+            // validate access rights and role permissions
+            ServiceResult serviceResult = await ValidatePermissionsAsync(
+                    context,
+                    nodeManager,
+                    sourceHandle,
+                    PermissionType.Browse,
+                    null,
+                    true,
+                    cancellationToken)
+                .ConfigureAwait(false);
+            if (ServiceResult.IsGood(serviceResult))
+            {
+                List<BrowsePathTarget> targets = [];
+                // translate path only if validation is passing
+                await TranslateBrowsePathAsync(
+                    context,
+                    nodeManager,
+                    sourceHandle,
+                    relativePath,
+                    targets,
+                    0,
+                    cancellationToken)
+                .ConfigureAwait(false);
+                result!.Targets = targets;
+            }
+
+            return serviceResult;
+        }
+
+        /// <summary>
+        /// Recursively processes the elements in the RelativePath starting at the specified index.
+        /// </summary>
+        /// <exception cref="ServiceResultException"></exception>
+        private async ValueTask TranslateBrowsePathAsync(
+            OperationContext context,
+            IAsyncNodeManager? nodeManager,
+            object? sourceHandle,
+            RelativePath relativePath,
+            List<BrowsePathTarget> targets,
+            int index,
+            CancellationToken cancellationToken)
+        {
+            Debug.Assert(nodeManager != null);
+            Debug.Assert(sourceHandle != null);
+            Debug.Assert(relativePath != null);
+            Debug.Assert(targets != null);
+
+            // check for end of list.
+            if (index < 0 || index >= relativePath!.Elements.Count)
+            {
+                return;
+            }
+
+            // follow the next hop.
+            RelativePathElement element = relativePath.Elements[index];
+
+            // check for valid reference type.
+            if (!element.IncludeSubtypes && element.ReferenceTypeId.IsNull)
+            {
+                return;
+            }
+
+            // check for valid target name.
+            if (element.TargetName.IsNull)
+            {
+                throw new ServiceResultException(StatusCodes.BadBrowseNameInvalid);
+            }
+
+            var targetIds = new List<ExpandedNodeId>();
+            var externalTargetIds = new List<NodeId>();
+
+            try
+            {
+                await nodeManager!.TranslateBrowsePathAsync(
+                    context,
+                    sourceHandle!,
+                    element,
+                    targetIds,
+                    externalTargetIds,
+                    cancellationToken)
+                .ConfigureAwait(false);
+            }
+            catch (Exception e)
+            {
+                m_logger.UnexpectedErrorTranslatingBrowsePath(e);
+                return;
+            }
+
+            // must check the browse name on all external targets.
+            for (int ii = 0; ii < externalTargetIds.Count; ii++)
+            {
+                // get the browse name from another node manager.
+                var description = new ReferenceDescription();
+
+                await UpdateReferenceDescriptionAsync(
+                        context,
+                        externalTargetIds[ii],
+                        NodeClass.Unspecified,
+                        BrowseResultMask.BrowseName,
+                        description,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+
+                // add to list if target name matches.
+                if (description.BrowseName == element.TargetName)
+                {
+                    bool found = false;
+
+                    for (int jj = 0; jj < targetIds.Count; jj++)
+                    {
+                        if (targetIds[jj] == externalTargetIds[ii])
+                        {
+                            found = true;
+                            break;
+                        }
+                    }
+
+                    if (!found)
+                    {
+                        targetIds.Add(externalTargetIds[ii]);
+                    }
+                }
+            }
+
+            // check if done after a final hop.
+            if (index == relativePath.Elements.Count - 1)
+            {
+                for (int ii = 0; ii < targetIds.Count; ii++)
+                {
+                    // Check the role permissions for target nodes
+                    (object? targetHandle, IAsyncNodeManager? targetNodeManager) =
+                        await m_owner.GetManagerHandleAsync(
+                            ExpandedNodeId.ToNodeId(targetIds[ii], Server.NamespaceUris),
+                            cancellationToken)
+                        .ConfigureAwait(false);
+
+                    if (targetHandle != null && targetNodeManager != null)
+                    {
+                        NodeMetadata nodeMetadata = await targetNodeManager.GetNodeMetadataAsync(
+                            context,
+                            targetHandle,
+                            BrowseResultMask.All,
+                            cancellationToken)
+                            .ConfigureAwait(false);
+
+                        ServiceResult serviceResult = MasterNodeManager.ValidateRolePermissions(
+                            context,
+                            nodeMetadata,
+                            PermissionType.Browse,
+                            m_logger);
+
+                        if (ServiceResult.IsBad(serviceResult))
+                        {
+                            // Remove target node without role permissions.
+                            continue;
+                        }
+                    }
+
+                    var target = new BrowsePathTarget
+                    {
+                        TargetId = targetIds[ii],
+                        RemainingPathIndex = uint.MaxValue
+                    };
+
+                    targets!.Add(target);
+                }
+
+                return;
+            }
+
+            // process next hops.
+            for (int ii = 0; ii < targetIds.Count; ii++)
+            {
+                ExpandedNodeId targetId = targetIds[ii];
+
+                // check for external reference.
+                if (targetId.IsAbsolute)
+                {
+                    var target = new BrowsePathTarget
+                    {
+                        TargetId = targetId,
+                        RemainingPathIndex = (uint)(index + 1)
+                    };
+
+                    targets!.Add(target);
+                    continue;
+                }
+
+                // check for valid start node.
+                (sourceHandle, nodeManager) = await m_owner.GetManagerHandleAsync((NodeId)targetId, cancellationToken)
+                    .ConfigureAwait(false);
+
+                if (sourceHandle == null)
+                {
+                    continue;
+                }
+
+                // recursively follow hops.
+                await TranslateBrowsePathAsync(
+                    context,
+                    nodeManager,
+                    sourceHandle,
+                    relativePath,
+                    targets!,
+                    index + 1,
+                    cancellationToken)
+                .ConfigureAwait(false);
+            }
+        }
+
+        internal async ValueTask<(ArrayOf<BrowseResult> results, ArrayOf<DiagnosticInfo> diagnosticInfos)> BrowseAsync(
+            OperationContext context,
+            ViewDescription view,
+            uint maxReferencesPerNode,
+            ArrayOf<BrowseDescription> nodesToBrowse,
+            CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            if (view != null && !view.ViewId.IsNull)
+            {
+                (object? viewHandle, IAsyncNodeManager? viewManager) =
+                    await m_owner.GetManagerHandleAsync(view.ViewId, cancellationToken)
+                    .ConfigureAwait(false);
+
+                if (viewHandle == null || viewManager == null)
+                {
+                    throw new ServiceResultException(StatusCodes.BadViewIdUnknown);
+                }
+
+                NodeMetadata metadata = await viewManager.GetNodeMetadataAsync(
+                        context,
+                        viewHandle,
+                        BrowseResultMask.NodeClass,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+
+                if (metadata == null || metadata.NodeClass != NodeClass.View)
+                {
+                    throw new ServiceResultException(StatusCodes.BadViewIdUnknown);
+                }
+
+                // validate access rights and role permissions
+                ServiceResult validationResult = await ValidatePermissionsAsync(
+                        context,
+                        viewManager,
+                        viewHandle,
+                        PermissionType.Browse,
+                        null,
+                        true,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+                if (ServiceResult.IsBad(validationResult))
+                {
+                    throw new ServiceResultException(validationResult);
+                }
+                view.Handle = viewHandle;
+            }
+
+            bool diagnosticsExist = false;
+            var results = new List<BrowseResult>(nodesToBrowse.Count);
+            var diagnosticInfos = new List<DiagnosticInfo>(nodesToBrowse.Count);
+
+            uint continuationPointsAssigned = 0;
+
+            for (int ii = 0; ii < nodesToBrowse.Count; ii++)
+            {
+                // check if request has timed out or been cancelled.
+                if (StatusCode.IsBad(context.OperationStatus))
+                {
+                    // release all allocated continuation points.
+                    foreach (BrowseResult current in results)
+                    {
+                        if (current != null && !current.ContinuationPoint.IsEmpty)
+                        {
+                            ContinuationPoint? cp = context.Session
+                                .ContinuationPoints.RestoreBrowse(current.ContinuationPoint);
+                            cp?.Dispose();
+                        }
+                    }
+
+                    throw new ServiceResultException(context.OperationStatus);
+                }
+
+                BrowseDescription nodeToBrowse = nodesToBrowse[ii];
+
+                // initialize result.
+                var result = new BrowseResult
+                {
+                    StatusCode = StatusCodes.Good,
+                    ContinuationPoint = default
+                };
+                results.Add(result);
+
+                ServiceResult error;
+
+                // need to trap unexpected exceptions to handle bugs in the node managers.
+                try
+                {
+                    error = await BrowseAsync(
+                        context,
+                        view,
+                        maxReferencesPerNode,
+                        continuationPointsAssigned < m_owner.MaxContinuationPointsPerBrowse,
+                        nodeToBrowse,
+                        result,
+                        cancellationToken).ConfigureAwait(false);
+                }
+                catch (Exception e)
+                {
+                    error = ServiceResult.Create(
+                        e,
+                        StatusCodes.BadUnexpectedError,
+                        "Unexpected error browsing node.");
+                }
+
+                // check for continuation point.
+                if (!result.ContinuationPoint.IsEmpty)
+                {
+                    continuationPointsAssigned++;
+                }
+
+                // check for error.
+                result.StatusCode = error.StatusCode;
+
+                if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                {
+                    DiagnosticInfo? diagnosticInfo = null;
+
+                    if (error != null && error.Code != StatusCodes.Good)
+                    {
+                        diagnosticInfo = ServerUtils.CreateDiagnosticInfo(Server, context, error, m_logger);
+                        diagnosticsExist = true;
+                    }
+
+                    diagnosticInfos.Add(diagnosticInfo!);
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (results, diagnosticInfos);
+        }
+
+        /// <summary>
+        /// Prepare a cache per NodeManager and unique NodeId that holds the attributes needed to validate the AccessRestrictions and RolePermissions.
+        /// This cache is then used in subsequenct calls to avoid triggering unnecessary time consuming callbacks.
+        /// The current services that benefit from this are the Read service
+        /// </summary>
+        /// <typeparam name="T">One of the following types used in the service calls:
+        ///     ReadValueId used in the Read service</typeparam>
+        /// <param name="nodesList">The collection of nodes on which the service operates uppon</param>
+        /// <param name="uniqueNodesServiceAttributes">The resulting cache that holds the values of the AccessRestrictions and RolePermissions attributes needed for Read service</param>
+        /// <exception cref="ArgumentException"></exception>
+        private static void PrepareValidationCache<T>(
+            ArrayOf<T> nodesList,
+            out Dictionary<NodeId, Variant[]> uniqueNodesServiceAttributes)
+        {
+            var uniqueNodes = new HashSet<NodeId>();
+            Type listType = typeof(T);
+
+            if (listType != typeof(ReadValueId) &&
+                listType != typeof(WriteValue) &&
+                listType != typeof(BrowseDescription) &&
+                listType != typeof(CallMethodRequest))
+            {
+                throw new ArgumentException(
+                    "Provided List<T> nodesCollection is of wrong type, T should be type BrowseDescription, ReadValueId or CallMethodRequest",
+                    nameof(nodesList));
+            }
+
+            for (int i = 0; i < nodesList.Count; i++)
+            {
+                NodeId nodeId = default;
+
+                if (listType == typeof(ReadValueId))
+                {
+                    nodeId = (nodesList[i] as ReadValueId)?.NodeId ?? default;
+                }
+                else if (listType == typeof(BrowseDescription))
+                {
+                    nodeId = (nodesList[i] as BrowseDescription)?.NodeId ?? default;
+                }
+                else if (listType == typeof(CallMethodRequest))
+                {
+                    nodeId = (nodesList[i] as CallMethodRequest)?.ObjectId ?? default;
+                }
+
+                if (!nodeId.IsNull)
+                {
+                    uniqueNodes.Add(nodeId);
+                }
+            }
+            // uniqueNodesReadAttributes is the place where the attributes for each unique nodeId are kept on the services
+            uniqueNodesServiceAttributes = [];
+            foreach (NodeId uniqueNode in uniqueNodes)
+            {
+                uniqueNodesServiceAttributes.Add(uniqueNode, []);
+            }
+        }
+
+        internal async ValueTask<(ArrayOf<BrowseResult> results, ArrayOf<DiagnosticInfo> diagnosticInfos)>
+            BrowseNextAsync(
+                OperationContext context,
+                bool releaseContinuationPoints,
+                ArrayOf<ByteString> continuationPoints,
+                CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            bool diagnosticsExist = false;
+            var results = new List<BrowseResult>(continuationPoints.Count);
+            var diagnosticInfos = new List<DiagnosticInfo>(continuationPoints.Count);
+
+            uint continuationPointsAssigned = 0;
+
+            for (int ii = 0; ii < continuationPoints.Count; ii++)
+            {
+                ContinuationPoint? cp;
+
+                // check if request has timed out or been canceled.
+                if (StatusCode.IsBad(context.OperationStatus))
+                {
+                    // release all allocated continuation points.
+                    foreach (BrowseResult current in results)
+                    {
+                        if (current != null && !current.ContinuationPoint.IsEmpty)
+                        {
+                            cp = context.Session
+                                .ContinuationPoints.RestoreBrowse(current.ContinuationPoint);
+                            cp?.Dispose();
+                        }
+                    }
+
+                    throw new ServiceResultException(context.OperationStatus);
+                }
+
+                // find the continuation point.
+                cp = context.Session.ContinuationPoints.RestoreBrowse(continuationPoints[ii]);
+
+                // validate access rights and role permissions
+                if (cp != null)
+                {
+                    ServiceResult validationResult = await ValidatePermissionsAsync(
+                            context,
+                            cp.Manager,
+                            cp.NodeToBrowse,
+                            PermissionType.Browse,
+                            null,
+                            true,
+                            cancellationToken)
+                        .ConfigureAwait(false);
+                    if (ServiceResult.IsBad(validationResult))
+                    {
+                        var badResult = new BrowseResult
+                        {
+                            StatusCode = validationResult.Code,
+                            ContinuationPoint = default
+                        };
+                        results.Add(badResult);
+
+                        // put placeholder for diagnostics
+                        diagnosticInfos.Add(null!);
+                        continue;
+                    }
+                }
+
+                // initialize result.
+                var result = new BrowseResult
+                {
+                    StatusCode = StatusCodes.Good,
+                    ContinuationPoint = default
+                };
+                results.Add(result);
+
+                // check if simply releasing the continuation point.
+                if (releaseContinuationPoints)
+                {
+                    cp?.Dispose();
+
+                    continue;
+                }
+
+                ServiceResult? error = null;
+
+                // check if continuation point has expired.
+                if (cp == null)
+                {
+                    error = StatusCodes.BadContinuationPointInvalid;
+                }
+
+                if (cp != null)
+                {
+                    // need to trap unexpected exceptions to handle bugs in the node managers.
+                    try
+                    {
+                        ArrayOf<ReferenceDescription> references = result.References;
+
+                        (error, cp, references) = await FetchReferencesAsync(
+                                context,
+                                continuationPointsAssigned < m_owner.MaxContinuationPointsPerBrowse,
+                                cp!,
+                                references,
+                                cancellationToken)
+                            .ConfigureAwait(false);
+
+                        result.References = references;
+                    }
+                    catch (Exception e)
+                    {
+                        error = ServiceResult.Create(
+                            e,
+                            StatusCodes.BadUnexpectedError,
+                            "Unexpected error browsing node.");
+                    }
+
+                    // check for continuation point.
+                    if (!result.ContinuationPoint.IsEmpty)
+                    {
+                        continuationPointsAssigned++;
+                    }
+                }
+
+                // check for error.
+                result.StatusCode = error!.StatusCode;
+
+                if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                {
+                    DiagnosticInfo? diagnosticInfo = null;
+
+                    if (error != null && error.Code != StatusCodes.Good)
+                    {
+                        diagnosticInfo = ServerUtils.CreateDiagnosticInfo(
+                            Server,
+                            context,
+                            error,
+                            m_logger);
+                        diagnosticsExist = true;
+                    }
+
+                    diagnosticInfos.Add(diagnosticInfo!);
+                }
+
+                // check for continuation point.
+                if (cp != null && ServiceResult.IsGood(error))
+                {
+                    result.StatusCode = StatusCodes.Good;
+                    result.ContinuationPoint = cp.Id.ToByteArray().ToByteString();
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (results, diagnosticInfos);
+        }
+
+        internal async ValueTask<ServiceResult> BrowseAsync(
+            OperationContext context,
+            ViewDescription? view,
+            uint maxReferencesPerNode,
+            bool assignContinuationPoint,
+            BrowseDescription nodeToBrowse,
+            BrowseResult result,
+            CancellationToken cancellationToken = default)
+        {
+            Debug.Assert(context != null);
+            Debug.Assert(nodeToBrowse != null);
+            Debug.Assert(result != null);
+
+            // find node manager that owns the node.
+            (object? handle, IAsyncNodeManager? nodeManager) = await m_owner.GetManagerHandleAsync(nodeToBrowse!.NodeId, cancellationToken)
+                .ConfigureAwait(false);
+
+            if (handle == null)
+            {
+                return StatusCodes.BadNodeIdUnknown;
+            }
+
+            if (!nodeToBrowse.ReferenceTypeId.IsNull &&
+                !Server.TypeTree.IsKnown(nodeToBrowse.ReferenceTypeId))
+            {
+                return StatusCodes.BadReferenceTypeIdInvalid;
+            }
+
+            if (nodeToBrowse.BrowseDirection is < BrowseDirection.Forward or > BrowseDirection.Both)
+            {
+                return StatusCodes.BadBrowseDirectionInvalid;
+            }
+
+            // validate access rights and role permissions
+            ServiceResult validationResult = await ValidatePermissionsAsync(
+                    context!,
+                    nodeManager,
+                    handle,
+                    PermissionType.Browse,
+                    null,
+                    true,
+                    cancellationToken)
+                .ConfigureAwait(false);
+            if (ServiceResult.IsBad(validationResult))
+            {
+                return validationResult;
+            }
+
+            // create a continuation point.
+            ContinuationPoint? tempCp = null;
+            try
+            {
+                tempCp = new ContinuationPoint
+                {
+                    Manager = nodeManager!,
+                    View = view,
+                    NodeToBrowse = handle,
+                    RequestedNodeId = nodeToBrowse.NodeId,
+                    MaxResultsToReturn = maxReferencesPerNode,
+                    BrowseDirection = nodeToBrowse.BrowseDirection,
+                    ReferenceTypeId = nodeToBrowse.ReferenceTypeId,
+                    IncludeSubtypes = nodeToBrowse.IncludeSubtypes,
+                    NodeClassMask = nodeToBrowse.NodeClassMask,
+                    ResultMask = (BrowseResultMask)nodeToBrowse.ResultMask,
+                    Index = 0,
+                    Data = null
+                };
+                ContinuationPoint? cp = tempCp;
+
+                // check if reference type left unspecified.
+                if (cp.ReferenceTypeId.IsNull)
+                {
+                    cp.ReferenceTypeId = ReferenceTypeIds.References;
+                    cp.IncludeSubtypes = true;
+                }
+
+                // loop until browse is complete or max results.
+                ArrayOf<ReferenceDescription> references = result!.References;
+
+                ServiceResult error;
+
+                (error, cp, references) = await FetchReferencesAsync(
+                   context!,
+                   assignContinuationPoint,
+                   cp!,
+                   references,
+                   cancellationToken)
+                   .ConfigureAwait(false);
+                tempCp = null; // ownership transferred to FetchReferencesAsync
+
+                result.References = references;
+
+                // save continuation point.
+                if (cp != null && ServiceResult.IsGood(error))
+                {
+                    result.StatusCode = StatusCodes.Good;
+                    result.ContinuationPoint = cp.Id.ToByteArray().ToByteString();
+                }
+
+                // all is good.
+                return error;
+            }
+            finally
+            {
+                tempCp?.Dispose();
+            }
+        }
+
+        internal async ValueTask<(
+            ServiceResult serviceResult,
+            ContinuationPoint? cp,
+            ArrayOf<ReferenceDescription> references
+            )> FetchReferencesAsync(
+                OperationContext context,
+                bool assignContinuationPoint,
+                ContinuationPoint cp,
+                ArrayOf<ReferenceDescription> references,
+                CancellationToken cancellationToken = default)
+        {
+            Debug.Assert(context != null);
+            Debug.Assert(cp != null);
+
+            IAsyncNodeManager nodeManager = cp!.Manager;
+            var nodeClassMask = (NodeClass)cp.NodeClassMask;
+            BrowseResultMask resultMask = cp.ResultMask;
+            var referenceList = references.ToList();
+            ContinuationPoint? currentCp = cp;
+            // loop until browse is complete or max results.
+            while (currentCp != null)
+            {
+                currentCp = await nodeManager.BrowseAsync(context!, currentCp, referenceList, cancellationToken)
+                    .ConfigureAwait(false);
+
+                var referencesToKeep = new List<ReferenceDescription>(referenceList.Count);
+
+                // check for incomplete reference descriptions.
+                for (int ii = 0; ii < referenceList.Count; ii++)
+                {
+                    ReferenceDescription reference = referenceList[ii];
+
+                    // check if filtering must be applied.
+                    if (reference.Unfiltered)
+                    {
+                        // ignore unknown external references.
+                        if (reference.NodeId.IsAbsolute)
+                        {
+                            continue;
+                        }
+
+                        // update the description.
+                        bool include = await UpdateReferenceDescriptionAsync(
+                                context!,
+                                (NodeId)reference.NodeId,
+                                nodeClassMask,
+                                resultMask,
+                                reference,
+                                cancellationToken)
+                            .ConfigureAwait(false);
+
+                        if (!include)
+                        {
+                            continue;
+                        }
+                    }
+
+                    // add to list.
+                    referencesToKeep.Add(reference);
+                }
+
+                // replace list.
+                referenceList = referencesToKeep;
+
+                // check if browse limit reached.
+                if (currentCp != null && referenceList.Count >= currentCp.MaxResultsToReturn)
+                {
+                    if (!assignContinuationPoint)
+                    {
+                        currentCp.Dispose();
+                        return (StatusCodes.BadNoContinuationPoints, null, referenceList);
+                    }
+
+                    currentCp.Id = Guid.NewGuid();
+                    context!.Session!.ContinuationPoints.SaveBrowse(currentCp);
+                    break;
+                }
+            }
+
+            // all is good.
+            return (ServiceResult.Good, currentCp, referenceList);
+        }
+
+        /// <summary>
+        /// Updates the reference description with the node attributes.
+        /// </summary>
+        /// <exception cref="ArgumentNullException"><paramref name="targetId"/> is <c>null</c>.</exception>
+        private async ValueTask<bool> UpdateReferenceDescriptionAsync(
+            OperationContext context,
+            NodeId targetId,
+            NodeClass nodeClassMask,
+            BrowseResultMask resultMask,
+            ReferenceDescription description,
+            CancellationToken cancellationToken = default)
+        {
+            if (targetId.IsNull)
+            {
+                throw new ArgumentNullException(nameof(targetId));
+            }
+
+            if (description == null)
+            {
+                throw new ArgumentNullException(nameof(description));
+            }
+
+            // find node manager that owns the node.
+            (object? handle, IAsyncNodeManager? nodeManager) = await m_owner.GetManagerHandleAsync(targetId, cancellationToken)
+                .ConfigureAwait(false);
+
+            // dangling reference - nothing more to do.
+            if (handle == null)
+            {
+                return false;
+            }
+
+            // fetch the node attributes.
+            NodeMetadata metadata = await nodeManager!.GetNodeMetadataAsync(context, handle, resultMask, cancellationToken)
+                .ConfigureAwait(false);
+
+            if (metadata == null)
+            {
+                return false;
+            }
+
+            // check nodeclass filter.
+            if (nodeClassMask != NodeClass.Unspecified &&
+                ((int)metadata.NodeClass & (int)nodeClassMask) == 0)
+            {
+                return false;
+            }
+
+            // update attributes.
+            description.NodeId = metadata.NodeId;
+
+            description.SetTargetAttributes(
+                resultMask,
+                metadata.NodeClass,
+                metadata.BrowseName,
+                metadata.DisplayName,
+                metadata.TypeDefinition);
+
+            description.Unfiltered = false;
+
+            return true;
+        }
+
+        internal async ValueTask<NodeState?> FindNodeInAddressSpaceAsync(NodeId nodeId, CancellationToken cancellationToken = default)
+        {
+            if (nodeId.IsNull)
+            {
+                return null;
+            }
+            // search node id in all node managers
+            foreach (IAsyncNodeManager nodeManager in m_owner.AsyncNodeManagers)
+            {
+                if ((await nodeManager.GetManagerHandleAsync(nodeId, cancellationToken).ConfigureAwait(false))
+                    is not NodeHandle handle)
+                {
+                    continue;
+                }
+                return handle.Node;
+            }
+            return null;
+        }
+
+        internal async ValueTask<(ArrayOf<DataValue> values, ArrayOf<DiagnosticInfo> diagnosticInfos)> ReadAsync(
+            OperationContext context,
+            double maxAge,
+            TimestampsToReturn timestampsToReturn,
+            ArrayOf<ReadValueId> nodesToRead,
+            CancellationToken cancellationToken = default)
+        {
+            if (maxAge < 0)
+            {
+                throw new ServiceResultException(StatusCodes.BadMaxAgeInvalid);
+            }
+
+            if (timestampsToReturn is < TimestampsToReturn.Source or > TimestampsToReturn.Neither)
+            {
+                throw new ServiceResultException(StatusCodes.BadTimestampsToReturnInvalid);
+            }
+
+            bool diagnosticsExist = false;
+            var values = new List<DataValue>(nodesToRead.Count);
+            var diagnosticInfos = new List<DiagnosticInfo>(nodesToRead.Count);
+
+            // create empty list of errors.
+            var errors = new List<ServiceResult>(values.Count);
+            for (int ii = 0; ii < nodesToRead.Count; ii++)
+            {
+                errors.Add(null!);
+            }
+
+            // add placeholder for each result.
+            bool validItems = false;
+
+            m_logger.MasterNodeManagerReadCountCount(nodesToRead.Count);
+
+            PrepareValidationCache(
+                nodesToRead,
+                out Dictionary<NodeId, Variant[]> uniqueNodesReadAttributes);
+
+            for (int ii = 0; ii < nodesToRead.Count; ii++)
+            {
+                // add default value to values collection
+                values.Add(default);
+                // add placeholder for diagnostics
+                diagnosticInfos.Add(null!);
+
+                // pre-validate and pre-parse parameter.
+                errors[ii] = (await ValidateReadRequestAsync(
+                    context,
+                    nodesToRead[ii],
+                    uniqueNodesReadAttributes,
+                    false,
+                    cancellationToken)
+                    .ConfigureAwait(false))!;
+
+                // return error status.
+                if (ServiceResult.IsBad(errors[ii]))
+                {
+                    nodesToRead[ii].Processed = true;
+                }
+                // found at least one valid item.
+                else
+                {
+                    nodesToRead[ii].Processed = false;
+                    validItems = true;
+                }
+            }
+
+            // call each node manager.
+            if (validItems)
+            {
+                foreach (IAsyncNodeManager nodeManager in m_nodeManagers)
+                {
+                    await nodeManager.ReadAsync(
+                        context,
+                        maxAge,
+                        nodesToRead,
+                        values,
+                        errors,
+                        cancellationToken).ConfigureAwait(false);
+                }
+            }
+
+            // process results.
+            for (int ii = 0; ii < nodesToRead.Count; ii++)
+            {
+                DataValue value = values[ii];
+
+                // set an error code for nodes that were not handled by any node manager.
+                if (!nodesToRead[ii].Processed)
+                {
+                    value = values[ii] = DataValue.FromStatusCode(
+                        StatusCodes.BadNodeIdUnknown,
+                        DateTime.UtcNow);
+                    errors[ii] = new ServiceResult(values[ii].StatusCode);
+                }
+
+                // update the diagnostic info and ensure the status code in the data value is the same as the error code.
+                if (errors[ii] != null && errors[ii].Code != StatusCodes.Good)
+                {
+                    if (value.IsNull)
+                    {
+                        value = values[ii] = DataValue.FromStatusCode(errors[ii].Code, DateTime.UtcNow);
+                    }
+
+                    value = values[ii] = value.WithStatus(errors[ii].Code);
+
+                    if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        diagnosticInfos[ii] = ServerUtils.CreateDiagnosticInfo(
+                            Server,
+                            context,
+                            errors[ii],
+                            m_logger)!;
+                        diagnosticsExist = true;
+                    }
+                }
+
+                // apply the timestamp filters.
+                if (timestampsToReturn is not TimestampsToReturn.Server and not TimestampsToReturn.Both)
+                {
+                    value = values[ii] = value.WithServerTimestamp(DateTimeUtc.MinValue);
+                }
+
+                if (timestampsToReturn is not TimestampsToReturn.Source and not TimestampsToReturn.Both)
+                {
+                    _ = values[ii] = value.WithSourceTimestamp(DateTimeUtc.MinValue);
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (values, diagnosticInfos);
+        }
+
+        internal async ValueTask<(ArrayOf<HistoryReadResult> values, ArrayOf<DiagnosticInfo> diagnosticInfos)> HistoryReadAsync(
+            OperationContext context,
+            ExtensionObject historyReadDetails,
+            TimestampsToReturn timestampsToReturn,
+            bool releaseContinuationPoints,
+            ArrayOf<HistoryReadValueId> nodesToRead,
+            CancellationToken cancellationToken = default)
+        {
+            // validate history details parameter.
+            if (historyReadDetails.IsNull)
+            {
+                throw new ServiceResultException(StatusCodes.BadHistoryOperationInvalid);
+            }
+
+            if (!historyReadDetails.TryGetValue(out HistoryReadDetails? details))
+            {
+                throw new ServiceResultException(StatusCodes.BadHistoryOperationInvalid);
+            }
+
+            // create result lists.
+            bool diagnosticsExist = false;
+            var results = new List<HistoryReadResult>(nodesToRead.Count);
+            var diagnosticInfos = new List<DiagnosticInfo>(nodesToRead.Count);
+
+            // pre-validate items.
+            bool validItems = false;
+            // create empty list of errors.
+            var errors = new List<ServiceResult>(results.Count);
+            for (int ii = 0; ii < nodesToRead.Count; ii++)
+            {
+                errors.Add(null!);
+            }
+
+            for (int ii = 0; ii < nodesToRead.Count; ii++)
+            {
+                // Limit permission restrictions to Client initiated service call
+                HistoryReadResult? result = null;
+                DiagnosticInfo? diagnosticInfo = null;
+
+                // pre-validate and pre-parse parameter.
+                errors[ii] = (await ValidateHistoryReadRequestAsync(context, nodesToRead[ii], cancellationToken)
+                    .ConfigureAwait(false))!;
+
+                // return error status.
+                if (ServiceResult.IsBad(errors[ii]))
+                {
+                    nodesToRead[ii].Processed = true;
+                    result = new HistoryReadResult { StatusCode = errors[ii].Code };
+
+                    // add diagnostics if requested.
+                    if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        diagnosticInfo = ServerUtils.CreateDiagnosticInfo(
+                            Server,
+                            context,
+                            errors[ii],
+                            m_logger);
+                        diagnosticsExist = true;
+                    }
+                }
+                // found at least one valid item.
+                else
+                {
+                    nodesToRead[ii].Processed = false;
+                    validItems = true;
+                }
+
+                results.Add(result!);
+                diagnosticInfos.Add(diagnosticInfo!);
+            }
+
+            // call each node manager.
+            if (validItems)
+            {
+                foreach (IAsyncNodeManager nodeManager in m_nodeManagers)
+                {
+                    await nodeManager.HistoryReadAsync(
+                         context,
+                        details!,
+                        timestampsToReturn,
+                        releaseContinuationPoints,
+                        nodesToRead,
+                        results,
+                        errors,
+                        cancellationToken).ConfigureAwait(false);
+                }
+
+                for (int ii = 0; ii < nodesToRead.Count; ii++)
+                {
+                    HistoryReadResult result = results[ii];
+
+                    // set an error code for nodes that were not handled by any node manager.
+                    if (!nodesToRead[ii].Processed)
+                    {
+                        nodesToRead[ii].Processed = true;
+                        result = results[ii] = new HistoryReadResult();
+                        result.StatusCode = StatusCodes.BadNodeIdUnknown;
+                        errors[ii] = results[ii].StatusCode;
+                    }
+
+                    // update the diagnostic info and ensure the status code in the result is the same as the error code.
+                    if (errors[ii] != null && errors[ii].Code != StatusCodes.Good)
+                    {
+                        result ??= results[ii] = new HistoryReadResult();
+
+                        result.StatusCode = errors[ii].Code;
+
+                        // add diagnostics if requested.
+                        if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                        {
+                            diagnosticInfos[ii] = ServerUtils.CreateDiagnosticInfo(
+                                Server,
+                                context,
+                                errors[ii],
+                                m_logger)!;
+                            diagnosticsExist = true;
+                        }
+                    }
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (results, diagnosticInfos);
+        }
+
+        internal async ValueTask<(ArrayOf<StatusCode> results, ArrayOf<DiagnosticInfo> diagnosticInfos)> WriteAsync(
+            OperationContext context,
+            ArrayOf<WriteValue> nodesToWrite,
+            CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            int count = nodesToWrite.Count;
+
+            bool diagnosticsExist = false;
+            var results = new List<StatusCode>(count);
+            var diagnosticInfos = new List<DiagnosticInfo>(count);
+
+            PrepareValidationCache(
+                nodesToWrite,
+                out Dictionary<NodeId, Variant[]> uniqueNodesReadAttributes);
+
+            // add placeholder for each result.
+            bool validItems = false;
+
+            for (int ii = 0; ii < count; ii++)
+            {
+                StatusCode result = StatusCodes.Good;
+                DiagnosticInfo? diagnosticInfo = null;
+
+                // pre-validate and pre-parse parameter. Validate also access rights and role permissions
+                ServiceResult? error = await ValidateWriteRequestAsync(
+                    context,
+                    nodesToWrite[ii],
+                    uniqueNodesReadAttributes,
+                    false,
+                    cancellationToken)
+                    .ConfigureAwait(false);
+
+                // return error status.
+                if (ServiceResult.IsBad(error))
+                {
+                    nodesToWrite[ii].Processed = true;
+                    result = error!.Code;
+
+                    // add diagnostics if requested.
+                    if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        diagnosticInfo = ServerUtils.CreateDiagnosticInfo(Server, context, error, m_logger);
+                        diagnosticsExist = true;
+                    }
+                }
+                // found at least one valid item.
+                else
+                {
+                    nodesToWrite[ii].Processed = false;
+                    validItems = true;
+                }
+
+                results.Add(result);
+                diagnosticInfos.Add(diagnosticInfo!);
+            }
+
+            // call each node manager.
+            if (validItems)
+            {
+                var errors = new List<ServiceResult>(count);
+                errors.AddRange(new ServiceResult[count]);
+
+                foreach (IAsyncNodeManager nodeManager in m_nodeManagers)
+                {
+                    await nodeManager.WriteAsync(
+                        context,
+                        nodesToWrite,
+                        errors,
+                        cancellationToken).ConfigureAwait(false);
+                }
+
+                for (int ii = 0; ii < nodesToWrite.Count; ii++)
+                {
+                    if (!nodesToWrite[ii].Processed)
+                    {
+                        errors[ii] = StatusCodes.BadNodeIdUnknown;
+                    }
+
+                    if (errors[ii] != null && errors[ii].Code != StatusCodes.Good)
+                    {
+                        results[ii] = errors[ii].Code;
+
+                        // add diagnostics if requested.
+                        if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                        {
+                            diagnosticInfos[ii] = ServerUtils.CreateDiagnosticInfo(
+                                Server,
+                                context,
+                                errors[ii],
+                                m_logger)!;
+                            diagnosticsExist = true;
+                        }
+                    }
+
+                    ServerUtils.ReportWriteValue(
+                        nodesToWrite[ii].NodeId,
+                        nodesToWrite[ii].Value,
+                        results[ii]);
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (results, diagnosticInfos);
+        }
+
+        internal async ValueTask<(ArrayOf<HistoryUpdateResult> results, ArrayOf<DiagnosticInfo> diagnosticInfos)>
+            HistoryUpdateAsync(
+                OperationContext context,
+                ArrayOf<ExtensionObject> historyUpdateDetails,
+                CancellationToken cancellationToken = default)
+        {
+            Type? detailsType = null;
+            var nodesToUpdate = new List<HistoryUpdateDetails>();
+
+            // verify that all extension objects in the list have the same type.
+            foreach (ExtensionObject details in historyUpdateDetails)
+            {
+                if (details.IsNull)
+                {
+                    continue;
+                }
+                if (!details.TryGetValue(out HistoryUpdateDetails? historyUpdateDetail))
+                {
+                    nodesToUpdate.Add(null!); // Retain old behavior
+                    continue;
+                }
+                detailsType = historyUpdateDetail!.GetType();
+                nodesToUpdate.Add(historyUpdateDetail);
+            }
+
+            // create result lists.
+            bool diagnosticsExist = false;
+            var results = new List<HistoryUpdateResult>(nodesToUpdate.Count);
+            var diagnosticInfos = new List<DiagnosticInfo>(nodesToUpdate.Count);
+
+            // pre-validate items.
+            bool validItems = false;
+
+            // create empty list of errors.
+            var errors = new List<ServiceResult>(results.Count);
+            for (int ii = 0; ii < nodesToUpdate.Count; ii++)
+            {
+                errors.Add(null!);
+            }
+
+            for (int ii = 0; ii < nodesToUpdate.Count; ii++)
+            {
+                HistoryUpdateResult? result = null;
+                DiagnosticInfo? diagnosticInfo = null;
+
+                // check the type of details parameter.
+                ServiceResult? error;
+                if (nodesToUpdate[ii].GetType() != detailsType)
+                {
+                    error = StatusCodes.BadHistoryOperationInvalid;
+                }
+                // pre-validate and pre-parse parameter.
+                else
+                {
+                    error = await ValidateHistoryUpdateRequestAsync(context, nodesToUpdate[ii], cancellationToken)
+                        .ConfigureAwait(false);
+                }
+
+                // return error status.
+                if (ServiceResult.IsBad(error))
+                {
+                    nodesToUpdate[ii].Processed = true;
+                    result = new HistoryUpdateResult { StatusCode = error!.Code };
+
+                    // add diagnostics if requested.
+                    if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        diagnosticInfo = ServerUtils.CreateDiagnosticInfo(Server, context, error, m_logger);
+                        diagnosticsExist = true;
+                    }
+                }
+                // found at least one valid item.
+                else
+                {
+                    nodesToUpdate[ii].Processed = false;
+                    validItems = true;
+                }
+
+                results.Add(result!);
+                diagnosticInfos.Add(diagnosticInfo!);
+            }
+
+            // call each node manager.
+            if (validItems)
+            {
+                foreach (IAsyncNodeManager nodeManager in m_nodeManagers)
+                {
+                    await nodeManager.HistoryUpdateAsync(
+                        context,
+                        detailsType!,
+                        nodesToUpdate,
+                        results,
+                        errors,
+                        cancellationToken).ConfigureAwait(false);
+                }
+
+                for (int ii = 0; ii < nodesToUpdate.Count; ii++)
+                {
+                    HistoryUpdateResult result = results[ii];
+
+                    // set an error code for nodes that were not handled by any node manager.
+                    if (!nodesToUpdate[ii].Processed)
+                    {
+                        nodesToUpdate[ii].Processed = true;
+                        result = results[ii] = new HistoryUpdateResult();
+                        result.StatusCode = StatusCodes.BadNodeIdUnknown;
+                        errors[ii] = result.StatusCode;
+                    }
+
+                    // update the diagnostic info and ensure the status code in the result is the same as the error code.
+                    if (errors[ii] != null && errors[ii].Code != StatusCodes.Good)
+                    {
+                        result ??= results[ii] = new HistoryUpdateResult();
+
+                        result.StatusCode = errors[ii].Code;
+
+                        // add diagnostics if requested.
+                        if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                        {
+                            diagnosticInfos[ii] = ServerUtils.CreateDiagnosticInfo(
+                                Server,
+                                context,
+                                errors[ii],
+                                m_logger)!;
+                            diagnosticsExist = true;
+                        }
+                    }
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (results, diagnosticInfos);
+        }
+
+        internal async ValueTask<(ArrayOf<CallMethodResult> results, ArrayOf<DiagnosticInfo> diagnosticInfos)>
+            CallAsync(
+                OperationContext context,
+                ArrayOf<CallMethodRequest> methodsToCall,
+                CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            bool diagnosticsExist = false;
+            var results = new List<CallMethodResult>(methodsToCall.Count);
+            var diagnosticInfos = new List<DiagnosticInfo>(methodsToCall.Count);
+            var errors = new List<ServiceResult>(methodsToCall.Count);
+
+            PrepareValidationCache(
+                methodsToCall,
+                out Dictionary<NodeId, Variant[]> uniqueNodesReadAttributes);
+
+            // add placeholder for each result.
+            bool validItems = false;
+
+            for (int ii = 0; ii < methodsToCall.Count; ii++)
+            {
+                results.Add(null!);
+                errors.Add(null!);
+
+                if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                {
+                    diagnosticInfos.Add(null!);
+                }
+
+                // validate request parameters.
+                errors[ii] = await ValidateCallRequestItemAsync(
+                    context,
+                    methodsToCall[ii],
+                    uniqueNodesReadAttributes,
+                    // With resolver-based method lookup and an attribute cache, GetPermissionMetadataAsync
+                    // reads and caches the required permission attributes without the permissionsOnly shortcut.
+                    false,
+                    cancellationToken).ConfigureAwait(false);
+
+                if (ServiceResult.IsBad(errors[ii]))
+                {
+                    methodsToCall[ii].Processed = true;
+
+                    // add diagnostics if requested.
+                    if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        diagnosticInfos[ii] = ServerUtils.CreateDiagnosticInfo(
+                            Server,
+                            context,
+                            errors[ii],
+                            m_logger)!;
+                        diagnosticsExist = true;
+                    }
+
+                    continue;
+                }
+
+                // found at least one valid item.
+                validItems = true;
+                methodsToCall[ii].Processed = false;
+            }
+
+            // call each node manager.
+            if (validItems)
+            {
+                foreach (IAsyncNodeManager nodeManager in m_nodeManagers)
+                {
+                    await nodeManager.CallAsync(
+                        context,
+                        methodsToCall,
+                        results,
+                        errors,
+                        cancellationToken).ConfigureAwait(false);
+                }
+            }
+
+            for (int ii = 0; ii < methodsToCall.Count; ii++)
+            {
+                // set an error code for calls that were not handled by any node manager.
+                if (!methodsToCall[ii].Processed)
+                {
+                    results[ii] = new CallMethodResult();
+                    errors[ii] = StatusCodes.BadNodeIdUnknown;
+                }
+
+                // update the diagnostic info and ensure the status code in the result is the same as the error code.
+                if (errors[ii] != null && errors[ii].Code != StatusCodes.Good)
+                {
+                    if (results[ii] == null)
+                    {
+                        results[ii] = new CallMethodResult();
+                    }
+
+                    results[ii].StatusCode = errors[ii].Code;
+
+                    // add diagnostics if requested.
+                    if ((context.DiagnosticsMask & DiagnosticsMasks.OperationAll) != 0)
+                    {
+                        diagnosticInfos[ii] = ServerUtils.CreateDiagnosticInfo(
+                            Server,
+                            context,
+                            errors[ii],
+                            m_logger)!;
+                        diagnosticsExist = true;
+                    }
+                }
+            }
+
+            // clear the diagnostics array if no diagnostics requested or no errors occurred.
+            UpdateDiagnostics(context, diagnosticsExist, ref diagnosticInfos);
+
+            return (results, diagnosticInfos);
+        }
+
+        internal async ValueTask ConditionRefreshAsync(
+            OperationContext context,
+            IList<IEventMonitoredItem> monitoredItems,
+            CancellationToken cancellationToken = default)
+        {
+            IAsyncNodeManager[] activeNodeManagers = [.. m_nodeManagers];
+            NotificationDispatchLease[] dispatches =
+                m_owner.GetConditionRefreshDispatches(
+                    activeNodeManagers,
+                    monitoredItems);
+            try
+            {
+                foreach (NotificationDispatchLease dispatch in dispatches)
+                {
+                    try
+                    {
+                        await dispatch.NodeManager
+                            .ConditionRefreshAsync(
+                                context,
+                                dispatch.MonitoredItems,
+                                cancellationToken)
+                            .ConfigureAwait(false);
+                    }
+                    catch (Exception e)
+                    {
+                        m_logger.ErrorCallingConditionRefreshAsyncOnAsyncNodeManager(e);
+                    }
+                }
+            }
+            finally
+            {
+                MasterNodeManager.DisposeNotificationDispatches(dispatches);
+            }
+        }
+
+        internal async ValueTask CreateMonitoredItemsAsync(
+            OperationContext context,
+            uint subscriptionId,
+            double publishingInterval,
+            TimestampsToReturn timestampsToReturn,
+            ArrayOf<MonitoredItemCreateRequest> itemsToCreate,
+            IList<ServiceResult> errors,
+            IList<MonitoringFilterResult> filterResults,
+            IList<IMonitoredItem> monitoredItems,
+            bool createDurable,
+            CancellationToken cancellationToken = default)
+        {
+            await CreateMonitoredItemsCoreAsync(
+                context,
+                subscriptionId,
+                publishingInterval,
+                timestampsToReturn,
+                itemsToCreate,
+                errors,
+                filterResults,
+                monitoredItems,
+                createDurable,
+                cancellationToken).ConfigureAwait(false);
+        }
+
+        private async ValueTask CreateMonitoredItemsCoreAsync(
+            OperationContext context,
+            uint subscriptionId,
+            double publishingInterval,
+            TimestampsToReturn timestampsToReturn,
+            ArrayOf<MonitoredItemCreateRequest> itemsToCreate,
+            IList<ServiceResult> errors,
+            IList<MonitoringFilterResult> filterResults,
+            IList<IMonitoredItem> monitoredItems,
+            bool createDurable,
+            CancellationToken cancellationToken)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            if (errors == null)
+            {
+                throw new ArgumentNullException(nameof(errors));
+            }
+
+            if (filterResults == null)
+            {
+                throw new ArgumentNullException(nameof(filterResults));
+            }
+
+            if (monitoredItems == null)
+            {
+                throw new ArgumentNullException(nameof(monitoredItems));
+            }
+
+            if (publishingInterval < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(publishingInterval));
+            }
+
+            if (timestampsToReturn is < TimestampsToReturn.Source or > TimestampsToReturn.Neither)
+            {
+                throw new ServiceResultException(StatusCodes.BadTimestampsToReturnInvalid);
+            }
+
+            // add placeholder for each result.
+            bool validItems = false;
+
+            for (int ii = 0; ii < itemsToCreate.Count; ii++)
+            {
+                // validate request parameters.
+                errors[ii] = (await ValidateMonitoredItemCreateRequestAsync(context, itemsToCreate[ii], cancellationToken)
+                    .ConfigureAwait(false))!;
+
+                if (ServiceResult.IsBad(errors[ii]))
+                {
+                    itemsToCreate[ii].Processed = true;
+                    continue;
+                }
+
+                // found at least one valid item.
+                validItems = true;
+                itemsToCreate[ii].Processed = false;
+            }
+
+            // call each node manager.
+            if (validItems)
+            {
+                // create items for event filters.
+                await CreateMonitoredItemsForEventsAsync(
+                        context,
+                        subscriptionId,
+                        publishingInterval,
+                        timestampsToReturn,
+                        itemsToCreate,
+                        errors,
+                        filterResults,
+                        monitoredItems,
+                        createDurable,
+                        m_monitoredItemIdFactory,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+
+                // create items for data access.
+                foreach (IAsyncNodeManager nodeManager in m_nodeManagers)
+                {
+                    await nodeManager.CreateMonitoredItemsAsync(
+                            context,
+                            subscriptionId,
+                            publishingInterval,
+                            timestampsToReturn,
+                            itemsToCreate,
+                            errors,
+                            filterResults,
+                            monitoredItems,
+                            createDurable,
+                            m_monitoredItemIdFactory,
+                            cancellationToken)
+                        .ConfigureAwait(false);
+                }
+
+                // fill results for unknown nodes.
+                for (int ii = 0; ii < errors.Count; ii++)
+                {
+                    if (!itemsToCreate[ii].Processed)
+                    {
+                        errors[ii] = new ServiceResult(StatusCodes.BadNodeIdUnknown);
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Create monitored items for event subscriptions.
+        /// </summary>
+        private async ValueTask CreateMonitoredItemsForEventsAsync(
+            OperationContext context,
+            uint subscriptionId,
+            double publishingInterval,
+            TimestampsToReturn timestampsToReturn,
+            ArrayOf<MonitoredItemCreateRequest> itemsToCreate,
+            IList<ServiceResult> errors,
+            IList<MonitoringFilterResult> filterResults,
+            IList<IMonitoredItem> monitoredItems,
+            bool createDurable,
+            MonitoredItemIdFactory monitoredItemIdFactory,
+            CancellationToken cancellationToken = default)
+        {
+            for (int ii = 0; ii < itemsToCreate.Count; ii++)
+            {
+                MonitoredItemCreateRequest itemToCreate = itemsToCreate[ii];
+
+                if (!itemToCreate.Processed)
+                {
+                    // all event subscriptions required an event filter.
+                    if (!itemToCreate.RequestedParameters.Filter.TryGetValue(out EventFilter? filter))
+                    {
+                        continue;
+                    }
+
+                    itemToCreate.Processed = true;
+
+                    // only the value attribute may be used with an event subscription.
+                    if (itemToCreate.ItemToMonitor.AttributeId != Attributes.EventNotifier)
+                    {
+                        errors[ii] = StatusCodes.BadFilterNotAllowed;
+                        continue;
+                    }
+
+                    // the index range parameter has no meaning for event subscriptions.
+                    if (!string.IsNullOrEmpty(itemToCreate.ItemToMonitor.IndexRange))
+                    {
+                        errors[ii] = StatusCodes.BadIndexRangeInvalid;
+                        continue;
+                    }
+
+                    // the data encoding has no meaning for event subscriptions.
+                    if (!itemToCreate.ItemToMonitor.DataEncoding.IsNull)
+                    {
+                        errors[ii] = StatusCodes.BadDataEncodingInvalid;
+                        continue;
+                    }
+
+                    // validate the event filter.
+                    EventFilter.Result result = filter!.Validate(
+                        new FilterContext(Server.NamespaceUris, Server.TypeTree, context, Server.Telemetry));
+
+                    if (ServiceResult.IsBad(result.Status))
+                    {
+                        errors[ii] = result.Status;
+                        filterResults[ii] = result.ToEventFilterResult(
+                            context.DiagnosticsMask,
+                            context.StringTable,
+                            m_logger);
+                        continue;
+                    }
+
+                    // check if a valid node.
+                    (object? handle, IAsyncNodeManager? nodeManager) = await m_owner.GetManagerHandleAsync(
+                        itemToCreate.ItemToMonitor.NodeId, cancellationToken)
+                        .ConfigureAwait(false);
+
+                    if (handle == null)
+                    {
+                        errors[ii] = StatusCodes.BadNodeIdUnknown;
+                        continue;
+                    }
+                    NodeMetadata nodeMetadata = await nodeManager!.GetNodeMetadataAsync(
+                            context,
+                            handle,
+                            BrowseResultMask.All,
+                            cancellationToken)
+                        .ConfigureAwait(false);
+
+                    errors[ii] = MasterNodeManager.ValidateRolePermissions(
+                        context,
+                        nodeMetadata,
+                        PermissionType.ReceiveEvents,
+                        m_logger);
+
+                    if (ServiceResult.IsBad(errors[ii]))
+                    {
+                        continue;
+                    }
+
+                    IEventMonitoredItem monitoredItem = Server.EventManager.CreateMonitoredItem(
+                        context,
+                        nodeManager,
+                        handle,
+                        subscriptionId,
+                        monitoredItemIdFactory,
+                        timestampsToReturn,
+                        publishingInterval,
+                        itemToCreate,
+                        filter,
+                        createDurable);
+
+                    // subscribe to all node managers.
+                    if (itemToCreate.ItemToMonitor.NodeId == Objects.Server)
+                    {
+                        foreach (IAsyncNodeManager manager in m_nodeManagers)
+                        {
+                            try
+                            {
+                                await manager.SubscribeToAllEventsAsync(
+                                    context,
+                                    subscriptionId,
+                                    monitoredItem,
+                                    false,
+                                    cancellationToken)
+                                    .ConfigureAwait(false);
+                            }
+                            catch (Exception e)
+                            {
+                                m_logger.NodeManagerThrewAnExceptionSubscribingToAll(e, manager.GetType().Name);
+                            }
+                        }
+                    }
+                    // only subscribe to the node manager that owns the node.
+                    else
+                    {
+                        ServiceResult error = await nodeManager.SubscribeToEventsAsync(
+                                context,
+                                handle,
+                                subscriptionId,
+                                monitoredItem,
+                                false,
+                                cancellationToken)
+                            .ConfigureAwait(false);
+
+                        if (ServiceResult.IsBad(error))
+                        {
+                            Server.EventManager.DeleteMonitoredItem(monitoredItem.Id);
+                            errors[ii] = error;
+                            continue;
+                        }
+                    }
+
+                    monitoredItems[ii] = monitoredItem;
+                    errors[ii] = StatusCodes.Good;
+                }
+            }
+        }
+
+        internal async ValueTask RestoreMonitoredItemsAsync(
+            IList<IStoredMonitoredItem> itemsToRestore,
+            IList<IMonitoredItem> monitoredItems,
+            IUserIdentity savedOwnerIdentity,
+            CancellationToken cancellationToken = default)
+        {
+            if (itemsToRestore == null)
+            {
+                throw new ArgumentNullException(nameof(itemsToRestore));
+            }
+
+            if (monitoredItems == null)
+            {
+                throw new ArgumentNullException(nameof(monitoredItems));
+            }
+
+            if (Server.IsRunning)
+            {
+                throw new InvalidOperationException(
+                    "Subscription restore can only occur on startup");
+            }
+
+            // Pre-hydrate monitored-item data/event queues from an asynchronous subscription
+            // store so a networked store (e.g. a high-availability mirror) can restore queued
+            // values on the synchronous monitored-item creation path
+            // (OPCFoundation/UA-.NETStandard#3939).
+            await PreHydrateMonitoredItemQueuesAsync(itemsToRestore, cancellationToken)
+                .ConfigureAwait(false);
+
+            // create items for event filters.
+            await RestoreMonitoredItemsForEventsAsync(itemsToRestore, monitoredItems, cancellationToken)
+                .ConfigureAwait(false);
+
+            // create items for data access.
+            foreach (IAsyncNodeManager nodeManager in m_nodeManagers)
+            {
+                await nodeManager.RestoreMonitoredItemsAsync(
+                        itemsToRestore,
+                        monitoredItems,
+                        savedOwnerIdentity,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+            }
+
+            for (int ii = 0; ii < itemsToRestore.Count; ii++)
+            {
+                IStoredMonitoredItem storedItem = itemsToRestore[ii];
+                if (storedItem.IsRestored ||
+                    (storedItem.TypeMask & MonitoredItemTypeMask.DataChange) == 0)
+                {
+                    continue;
+                }
+
+                var monitoredItem = new MonitoredItem(
+                    Server,
+                    MonitoredItem.GetDetachedOwner(Server),
+                    MonitoredItem.DetachedHandle,
+                    storedItem);
+                var lifecycle = (IDetachableMonitoredItem)monitoredItem;
+                lifecycle.Detach(Server);
+                lifecycle.MarkNodeDeleted();
+                storedItem.IsRestored = true;
+                monitoredItems[ii] = monitoredItem;
+            }
+
+            m_monitoredItemIdFactory.SetStartValue(itemsToRestore.Max(i => i.Id));
+        }
+
+        internal async ValueTask PreHydrateMonitoredItemQueuesAsync(
+            IList<IStoredMonitoredItem> itemsToRestore,
+            CancellationToken cancellationToken)
+        {
+            ISubscriptionStore? subscriptionStore = Server.SubscriptionStore;
+            if (subscriptionStore == null)
+            {
+                return;
+            }
+
+            foreach (IStoredMonitoredItem item in itemsToRestore)
+            {
+                try
+                {
+                    if (item.TypeMask == MonitoredItemTypeMask.DataChange)
+                    {
+                        if (item.QueueSize > 1 || item.IsDurable)
+                        {
+                            item.RestoredDataChangeQueue = await subscriptionStore
+                                .RestoreDataChangeMonitoredItemQueueAsync(item.Id, cancellationToken)
+                                .ConfigureAwait(false);
+                        }
+                    }
+                    else if ((item.TypeMask & MonitoredItemTypeMask.Events) != 0)
+                    {
+                        item.RestoredEventQueue = await subscriptionStore
+                            .RestoreEventMonitoredItemQueueAsync(item.Id, cancellationToken)
+                            .ConfigureAwait(false);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    m_logger.FailedToPreHydrateQueueForMonitored(ex, item.Id);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Restore monitored items for event subscriptions.
+        /// </summary>
+        private async ValueTask RestoreMonitoredItemsForEventsAsync(
+            IList<IStoredMonitoredItem> itemsToRestore,
+            IList<IMonitoredItem> monitoredItems,
+            CancellationToken cancellationToken = default)
+        {
+            for (int ii = 0; ii < itemsToRestore.Count; ii++)
+            {
+                IStoredMonitoredItem item = itemsToRestore[ii];
+
+                if (!item.IsRestored)
+                {
+                    // all event subscriptions required an event filter.
+                    if (item.OriginalFilter is not EventFilter)
+                    {
+                        continue;
+                    }
+
+                    item.IsRestored = true;
+
+                    // check if a valid node.
+                    (object? handle, IAsyncNodeManager? nodeManager) = await m_owner.GetManagerHandleAsync(item.NodeId, cancellationToken)
+                        .ConfigureAwait(false);
+
+                    if (handle == null)
+                    {
+                        continue;
+                    }
+
+                    IEventMonitoredItem monitoredItem = Server.EventManager.RestoreMonitoredItem(
+                        nodeManager!,
+                        handle,
+                        item);
+
+                    // subscribe to all node managers.
+                    if (item.NodeId == Objects.Server)
+                    {
+                        foreach (IAsyncNodeManager manager in m_nodeManagers)
+                        {
+                            try
+                            {
+                                using var eventContext = new OperationContext(monitoredItem);
+                                await manager.SubscribeToAllEventsAsync(
+                                        eventContext,
+                                        monitoredItem.SubscriptionId,
+                                        monitoredItem,
+                                        false,
+                                        cancellationToken)
+                                    .ConfigureAwait(false);
+                            }
+                            catch (Exception e)
+                            {
+                                m_logger.NodeManagerThrewAnExceptionSubscribingToAll(e, manager.GetType().Name);
+                            }
+                        }
+                    }
+                    // only subscribe to the node manager that owns the node.
+                    else
+                    {
+                        using var eventContext = new OperationContext(monitoredItem);
+                        ServiceResult error = await nodeManager!.SubscribeToEventsAsync(
+                                eventContext,
+                                handle,
+                                monitoredItem.SubscriptionId,
+                                monitoredItem,
+                                false,
+                                cancellationToken)
+                            .ConfigureAwait(false);
+
+                        if (ServiceResult.IsBad(error))
+                        {
+                            Server.EventManager.DeleteMonitoredItem(monitoredItem.Id);
+                            continue;
+                        }
+                    }
+
+                    monitoredItems[ii] = monitoredItem;
+                }
+            }
+        }
+
+        internal async ValueTask ModifyMonitoredItemsAsync(
+            OperationContext context,
+            TimestampsToReturn timestampsToReturn,
+            IList<IMonitoredItem> monitoredItems,
+            ArrayOf<MonitoredItemModifyRequest> itemsToModify,
+            IList<ServiceResult> errors,
+            IList<MonitoringFilterResult> filterResults,
+            CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            if (monitoredItems == null)
+            {
+                throw new ArgumentNullException(nameof(monitoredItems));
+            }
+
+            if (errors == null)
+            {
+                throw new ArgumentNullException(nameof(errors));
+            }
+
+            if (filterResults == null)
+            {
+                throw new ArgumentNullException(nameof(filterResults));
+            }
+
+            if (timestampsToReturn is < TimestampsToReturn.Source or > TimestampsToReturn.Neither)
+            {
+                throw new ServiceResultException(StatusCodes.BadTimestampsToReturnInvalid);
+            }
+
+            bool validItems = false;
+
+            for (int ii = 0; ii < itemsToModify.Count; ii++)
+            {
+                // check for errors.
+                if (ServiceResult.IsBad(errors[ii]) || monitoredItems[ii] == null)
+                {
+                    itemsToModify[ii].Processed = true;
+                    continue;
+                }
+
+                if (monitoredItems[ii] is IDetachableMonitoredItem
+                    {
+                        IsDetached: true
+                    })
+                {
+                    errors[ii] = StatusCodes.BadNodeIdUnknown;
+                    itemsToModify[ii].Processed = true;
+                    continue;
+                }
+
+                // validate request parameters.
+                errors[ii] = MasterNodeManager.ValidateMonitoredItemModifyRequest(itemsToModify[ii])!;
+
+                if (ServiceResult.IsBad(errors[ii]))
+                {
+                    itemsToModify[ii].Processed = true;
+                    continue;
+                }
+
+                // found at least one valid item.
+                validItems = true;
+                itemsToModify[ii].Processed = false;
+            }
+
+            // call each node manager.
+            if (validItems)
+            {
+                // modify items for event filters.
+                await ModifyMonitoredItemsForEventsAsync(
+                        context,
+                        timestampsToReturn,
+                        monitoredItems,
+                        itemsToModify,
+                        errors,
+                        filterResults,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+
+                // let each owning node manager modify the items it created. Data monitored
+                // items are dispatched to their recorded owning NodeManager (grouped by
+                // owner) rather than to visible routing-table managers only, so items still
+                // owned by a shadow-retired generation are handled by that generation.
+                await DispatchModifyToOwningNodeManagersAsync(
+                        context,
+                        timestampsToReturn,
+                        monitoredItems,
+                        itemsToModify,
+                        errors,
+                        filterResults,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+
+                // update results.
+                for (int ii = 0; ii < errors.Count; ii++)
+                {
+                    if (!itemsToModify[ii].Processed)
+                    {
+                        errors[ii] = new ServiceResult(StatusCodes.BadMonitoredItemIdInvalid);
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Modify monitored items for event subscriptions.
+        /// </summary>
+        private async ValueTask ModifyMonitoredItemsForEventsAsync(
+            OperationContext context,
+            TimestampsToReturn timestampsToReturn,
+            IList<IMonitoredItem> monitoredItems,
+            ArrayOf<MonitoredItemModifyRequest> itemsToModify,
+            IList<ServiceResult> errors,
+            IList<MonitoringFilterResult> filterResults,
+            CancellationToken cancellationToken = default)
+        {
+            for (int ii = 0; ii < itemsToModify.Count; ii++)
+            {
+                if (itemsToModify[ii].Processed)
+                {
+                    continue;
+                }
+
+                // all event subscriptions are handled by the event manager.
+                if (monitoredItems[ii] is not IEventMonitoredItem monitoredItem ||
+                    (monitoredItem.MonitoredItemType & MonitoredItemTypeMask.Events) == 0)
+                {
+                    continue;
+                }
+
+                MonitoredItemModifyRequest itemToModify = itemsToModify[ii];
+                itemToModify.Processed = true;
+
+                // check for a valid filter.
+                if (itemToModify.RequestedParameters.Filter.IsNull)
+                {
+                    errors[ii] = StatusCodes.BadEventFilterInvalid;
+                    continue;
+                }
+
+                // all event subscriptions required an event filter.
+
+                if (!itemToModify.RequestedParameters.Filter.TryGetValue(out EventFilter? filter))
+                {
+                    errors[ii] = StatusCodes.BadEventFilterInvalid;
+                    continue;
+                }
+
+                // validate the event filter.
+                EventFilter.Result result = filter!.Validate(
+                    new FilterContext(Server.NamespaceUris, Server.TypeTree, context, Server.Telemetry));
+
+                if (ServiceResult.IsBad(result.Status))
+                {
+                    errors[ii] = result.Status;
+                    filterResults[ii] = result.ToEventFilterResult(
+                        context.DiagnosticsMask,
+                        context.StringTable,
+                        m_logger);
+                    continue;
+                }
+
+                // modify the item.
+                Server.EventManager.ModifyMonitoredItem(
+                    context,
+                    monitoredItem,
+                    timestampsToReturn,
+                    itemToModify,
+                    filter);
+
+                // subscribe to all node managers.
+                if ((monitoredItem.MonitoredItemType & MonitoredItemTypeMask.AllEvents) != 0)
+                {
+                    IAsyncNodeManager[] activeNodeManagers = [.. m_nodeManagers];
+                    NotificationDispatchLease[] dispatches =
+                        m_owner.GetAllEventNotificationDispatches(
+                            activeNodeManagers,
+                            monitoredItem);
+                    try
+                    {
+                        foreach (NotificationDispatchLease dispatch in dispatches)
+                        {
+                            await dispatch.NodeManager.SubscribeToAllEventsAsync(
+                                    context,
+                                    monitoredItem.SubscriptionId,
+                                    monitoredItem,
+                                    false,
+                                    cancellationToken)
+                                .ConfigureAwait(false);
+                        }
+                    }
+                    finally
+                    {
+                        MasterNodeManager.DisposeNotificationDispatches(dispatches);
+                    }
+                }
+                // only subscribe to the node manager that owns the node.
+                else
+                {
+                    await monitoredItem.NodeManager.SubscribeToEventsAsync(
+                        context,
+                        monitoredItem.ManagerHandle,
+                        monitoredItem.SubscriptionId,
+                        monitoredItem,
+                        false,
+                        cancellationToken).ConfigureAwait(false);
+                }
+
+                errors[ii] = StatusCodes.Good;
+            }
+        }
+
+        internal async ValueTask TransferMonitoredItemsAsync(
+            OperationContext context,
+            bool sendInitialValues,
+            IList<IMonitoredItem> monitoredItems,
+            IList<ServiceResult> errors,
+            MonitoredItemTransferOptions transferOptions,
+            CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+            if (monitoredItems == null)
+            {
+                throw new ArgumentNullException(nameof(monitoredItems));
+            }
+            if (errors == null)
+            {
+                throw new ArgumentNullException(nameof(errors));
+            }
+
+            IMonitoredItemTransferTransaction transaction =
+                await PrepareMonitoredItemsTransferAsync(
+                    context,
+                    sendInitialValues,
+                    monitoredItems,
+                    errors,
+                    transferOptions,
+                    cancellationToken).ConfigureAwait(false);
+            transaction.Commit();
+        }
+
+        internal async ValueTask<IMonitoredItemTransferTransaction>
+            PrepareMonitoredItemsTransferAsync(
+                OperationContext destinationContext,
+                bool sendInitialValues,
+                IList<IMonitoredItem> monitoredItems,
+                IList<ServiceResult> errors,
+                MonitoredItemTransferOptions transferOptions,
+                CancellationToken cancellationToken)
+        {
+            if (destinationContext == null)
+            {
+                throw new ArgumentNullException(nameof(destinationContext));
+            }
+            if (monitoredItems == null)
+            {
+                throw new ArgumentNullException(nameof(monitoredItems));
+            }
+            if (errors == null)
+            {
+                throw new ArgumentNullException(nameof(errors));
+            }
+
+            var processedItems = new List<bool>(monitoredItems.Count);
+            var effectiveTransferOptions = new MonitoredItemTransferOptions
+            {
+                DeferInitialValues = sendInitialValues ||
+                    transferOptions.DeferInitialValues
+            };
+
+            // preset results for unknown nodes
+            for (int ii = 0; ii < monitoredItems.Count; ii++)
+            {
+                IMonitoredItem? monitoredItem = monitoredItems[ii];
+                bool isDetached = monitoredItem is IDetachableMonitoredItem
+                {
+                    IsDetached: true
+                };
+                processedItems.Add(
+                    monitoredItem == null ||
+                    isDetached);
+                errors[ii] = isDetached
+                    ? ServiceResult.Good
+                    : new ServiceResult(StatusCodes.BadMonitoredItemIdInvalid);
+            }
+
+            List<(IAsyncNodeManager Owner, List<int> Indices)>? owners =
+                GroupDataMonitoredItemsByOwner(
+                    monitoredItems,
+                    index => processedItems[index]);
+            if (owners is not null)
+            {
+                foreach ((IAsyncNodeManager owner, List<int> indices) in owners)
+                {
+                    // Present only this owner's items as unprocessed.
+                    bool[] ownedItems = new bool[monitoredItems.Count];
+                    for (int ii = 0; ii < ownedItems.Length; ii++)
+                    {
+                        ownedItems[ii] = true;
+                    }
+                    foreach (int ii in indices)
+                    {
+                        ownedItems[ii] = false;
+                    }
+
+                    try
+                    {
+                        await owner.TransferMonitoredItemsAsync(
+                                destinationContext,
+                                sendInitialValues,
+                                monitoredItems,
+                                ownedItems,
+                                errors,
+                                effectiveTransferOptions,
+                                cancellationToken)
+                            .ConfigureAwait(false);
+                    }
+                    catch (Exception transferError)
+                    {
+                        m_logger.MonitoredItemTransferFailedForNodeManager(
+                            transferError,
+                            owner.GetType().Name);
+                        foreach (int ii in indices)
+                        {
+                            errors[ii] = ServiceResult.Create(
+                                transferError,
+                                StatusCodes.BadUnexpectedError,
+                                string.Empty);
+                            ownedItems[ii] = true;
+                            processedItems[ii] = true;
+                        }
+                        continue;
+                    }
+
+                    // Merge the owner's processed marks back into the shared list.
+                    foreach (int ii in indices)
+                    {
+                        if (ownedItems[ii])
+                        {
+                            processedItems[ii] = true;
+                        }
+                    }
+                }
+            }
+
+            return new MonitoredItemTransferTransaction(
+                sendInitialValues,
+                monitoredItems,
+                errors);
+        }
+
+        private sealed class MonitoredItemTransferTransaction :
+            IMonitoredItemTransferTransaction
+        {
+            public MonitoredItemTransferTransaction(
+                bool sendInitialValues,
+                IList<IMonitoredItem> monitoredItems,
+                IList<ServiceResult> errors)
+            {
+                m_sendInitialValues = sendInitialValues;
+                m_monitoredItems = monitoredItems;
+                m_errors = errors;
+            }
+
+            public void Commit()
+            {
+                if (Interlocked.CompareExchange(ref m_state, 1, 0) != 0)
+                {
+                    throw new InvalidOperationException("The monitored-item transfer is no longer pending.");
+                }
+
+                if (!m_sendInitialValues)
+                {
+                    return;
+                }
+
+                for (int ii = 0; ii < m_monitoredItems.Count; ii++)
+                {
+                    IMonitoredItem? monitoredItem = m_monitoredItems[ii];
+                    if (monitoredItem == null || ServiceResult.IsBad(m_errors[ii]))
+                    {
+                        continue;
+                    }
+                    if (monitoredItem is IDetachableMonitoredItem
+                        {
+                            IsDetached: true
+                        } detachedMonitoredItem)
+                    {
+                        detachedMonitoredItem.QueueNodeIdUnknown();
+                    }
+                    else
+                    {
+                        monitoredItem.SetupResendDataTrigger();
+                    }
+                }
+            }
+
+            private readonly bool m_sendInitialValues;
+            private readonly IList<IMonitoredItem> m_monitoredItems;
+            private readonly IList<ServiceResult> m_errors;
+            private int m_state;
+        }
+
+        internal async ValueTask DeleteMonitoredItemsAsync(
+            OperationContext context,
+            uint subscriptionId,
+            IList<IMonitoredItem> itemsToDelete,
+            IList<ServiceResult> errors,
+            CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            if (itemsToDelete == null)
+            {
+                throw new ArgumentNullException(nameof(itemsToDelete));
+            }
+
+            if (errors == null)
+            {
+                throw new ArgumentNullException(nameof(errors));
+            }
+
+            var processedItems = new List<bool>(itemsToDelete.Count);
+
+            for (int ii = 0; ii < itemsToDelete.Count; ii++)
+            {
+                IMonitoredItem? monitoredItem = itemsToDelete[ii];
+                bool isDetached = monitoredItem is IDetachableMonitoredItem
+                {
+                    IsDetached: true
+                };
+                processedItems.Add(
+                    ServiceResult.IsBad(errors[ii]) ||
+                    monitoredItem == null ||
+                    isDetached);
+                if (isDetached && monitoredItem is not null)
+                {
+                    errors[ii] = ServiceResult.Good;
+                    if ((monitoredItem.MonitoredItemType & MonitoredItemTypeMask.Events) != 0)
+                    {
+                        Server.EventManager.DeleteMonitoredItem(monitoredItem.Id);
+                    }
+                }
+            }
+
+            // delete items for event filters.
+            await DeleteMonitoredItemsForEventsAsync(
+                    context,
+                    subscriptionId,
+                    itemsToDelete,
+                    processedItems,
+                    errors,
+                    cancellationToken)
+                .ConfigureAwait(false);
+
+            // call each owning node manager. Data monitored items are dispatched to their
+            // recorded owning NodeManager (grouped by owner) so items owned by a
+            // shadow-retired generation are deleted by that generation, draining it.
+            await DispatchDataMonitoredItemsToOwningNodeManagersAsync(
+                    itemsToDelete,
+                    processedItems,
+                    (owner, ownedItems) => owner.DeleteMonitoredItemsAsync(
+                        context,
+                        itemsToDelete,
+                        ownedItems,
+                        errors,
+                        cancellationToken),
+                    notifyRetiredGenerationDrain: true,
+                    cancellationToken)
+                .ConfigureAwait(false);
+
+            // fill results for unknown nodes.
+            for (int ii = 0; ii < errors.Count; ii++)
+            {
+                if (!processedItems[ii])
+                {
+                    errors[ii] = StatusCodes.BadMonitoredItemIdInvalid;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Delete monitored items for event subscriptions.
+        /// </summary>
+        private async ValueTask DeleteMonitoredItemsForEventsAsync(
+            OperationContext context,
+            uint subscriptionId,
+            IList<IMonitoredItem> monitoredItems,
+            List<bool> processedItems,
+            IList<ServiceResult> errors,
+            CancellationToken cancellationToken = default)
+        {
+            bool retiredGenerationDrained = false;
+            for (int ii = 0; ii < monitoredItems.Count; ii++)
+            {
+                if (processedItems[ii])
+                {
+                    continue;
+                }
+
+                // all event subscriptions are handled by the event manager.
+                if (monitoredItems[ii] is not IEventMonitoredItem monitoredItem ||
+                    (monitoredItem.MonitoredItemType & MonitoredItemTypeMask.Events) == 0)
+                {
+                    continue;
+                }
+
+                IAsyncNodeManager owningNodeManager = monitoredItem.NodeManager;
+                processedItems[ii] = true;
+
+                // unsubscribe to all node managers.
+                if ((monitoredItem.MonitoredItemType & MonitoredItemTypeMask.AllEvents) != 0)
+                {
+                    NotificationDispatchLease[] dispatches =
+                        m_owner.GetAllEventUnsubscribeDispatches(monitoredItem);
+                    try
+                    {
+                        foreach (NotificationDispatchLease dispatch in dispatches)
+                        {
+                            await dispatch.NodeManager.SubscribeToAllEventsAsync(
+                                    context,
+                                    subscriptionId,
+                                    monitoredItem,
+                                    true,
+                                    cancellationToken)
+                                .ConfigureAwait(false);
+                            if (dispatch.Notifications is not null)
+                            {
+                                m_owner.CompleteRetiredAllEventUnsubscribe(
+                                    monitoredItem,
+                                    dispatch.Notifications);
+                            }
+                        }
+                        m_owner.CompleteRetiredAllEventUnsubscribe(
+                            monitoredItem,
+                            dispatches);
+                    }
+                    finally
+                    {
+                        MasterNodeManager.DisposeNotificationDispatches(dispatches);
+                    }
+                }
+                // only unsubscribe to the node manager that owns the node.
+                else
+                {
+                    await owningNodeManager.SubscribeToEventsAsync(
+                        context,
+                        monitoredItem.ManagerHandle,
+                        subscriptionId,
+                        monitoredItem,
+                        true,
+                        cancellationToken).ConfigureAwait(false);
+                }
+
+                // delete the item.
+                Server.EventManager.DeleteMonitoredItem(monitoredItem.Id);
+                retiredGenerationDrained |= !m_nodeManagers.Contains(owningNodeManager);
+
+                // success.
+                errors[ii] = StatusCodes.Good;
+            }
+
+            if (retiredGenerationDrained)
+            {
+                m_owner.NotifyRetiredGenerationDrainObserver();
+            }
+        }
+
+        internal async ValueTask SetMonitoringModeAsync(
+            OperationContext context,
+            MonitoringMode monitoringMode,
+            IList<IMonitoredItem> itemsToModify,
+            IList<ServiceResult> errors,
+            CancellationToken cancellationToken = default)
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            if (itemsToModify == null)
+            {
+                throw new ArgumentNullException(nameof(itemsToModify));
+            }
+
+            if (errors == null)
+            {
+                throw new ArgumentNullException(nameof(errors));
+            }
+
+            // call each node manager.
+            var processedItems = new List<bool>(itemsToModify.Count);
+
+            for (int ii = 0; ii < itemsToModify.Count; ii++)
+            {
+                IMonitoredItem? monitoredItem = itemsToModify[ii];
+                bool isDetached = monitoredItem is IDetachableMonitoredItem
+                {
+                    IsDetached: true
+                };
+                processedItems.Add(
+                    ServiceResult.IsBad(errors[ii]) ||
+                    monitoredItem == null ||
+                    isDetached);
+                if (isDetached && monitoredItem is not null)
+                {
+                    MonitoringMode previousMode =
+                        monitoredItem.SetMonitoringMode(monitoringMode);
+                    if (monitoringMode == MonitoringMode.Reporting &&
+                        previousMode != MonitoringMode.Reporting)
+                    {
+                        ((IDetachableMonitoredItem)monitoredItem).QueueNodeIdUnknown();
+                    }
+                    errors[ii] = ServiceResult.Good;
+                }
+            }
+
+            // delete items for event filters.
+            SetMonitoringModeForEvents(
+                context,
+                monitoringMode,
+                itemsToModify,
+                processedItems,
+                errors);
+
+            // set the monitoring mode on each owning node manager. Data monitored items are
+            // dispatched to their recorded owning NodeManager (grouped by owner) so items
+            // owned by a shadow-retired generation are handled by that generation.
+            await DispatchDataMonitoredItemsToOwningNodeManagersAsync(
+                    itemsToModify,
+                    processedItems,
+                    (owner, ownedItems) => owner.SetMonitoringModeAsync(
+                        context,
+                        monitoringMode,
+                        itemsToModify,
+                        ownedItems,
+                        errors,
+                        cancellationToken),
+                    notifyRetiredGenerationDrain: false,
+                    cancellationToken)
+                .ConfigureAwait(false);
+
+            // fill results for unknown nodes.
+            for (int ii = 0; ii < errors.Count; ii++)
+            {
+                if (!processedItems[ii])
+                {
+                    errors[ii] = StatusCodes.BadMonitoredItemIdInvalid;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Delete monitored items for event subscriptions.
+        /// </summary>
+        private static void SetMonitoringModeForEvents(
+            OperationContext context,
+            MonitoringMode monitoringMode,
+            IList<IMonitoredItem> monitoredItems,
+            List<bool> processedItems,
+            IList<ServiceResult> errors)
+        {
+            for (int ii = 0; ii < monitoredItems.Count; ii++)
+            {
+                if (processedItems[ii])
+                {
+                    continue;
+                }
+
+                // all event subscriptions are handled by the event manager.
+                if (monitoredItems[ii] is not IEventMonitoredItem monitoredItem ||
+                    (monitoredItem.MonitoredItemType & MonitoredItemTypeMask.Events) == 0)
+                {
+                    continue;
+                }
+
+                processedItems[ii] = true;
+
+                // set the monitoring mode.
+                monitoredItem.SetMonitoringMode(monitoringMode);
+
+                // success.
+                errors[ii] = StatusCodes.Good;
+            }
+        }
+
+        /// <summary>
+        /// Groups the unprocessed data monitored items by their recorded owning
+        /// NodeManager (by reference), preserving the original index of each item. Event
+        /// items and items already handled (processed) or unknown (null) are skipped.
+        /// </summary>
+        private static List<(IAsyncNodeManager Owner, List<int> Indices)>?
+            GroupDataMonitoredItemsByOwner(
+                IList<IMonitoredItem> monitoredItems,
+                Func<int, bool> isProcessed)
+        {
+            List<(IAsyncNodeManager Owner, List<int> Indices)>? owners = null;
+            Dictionary<object, int>? ownerIndex = null;
+            for (int ii = 0; ii < monitoredItems.Count; ii++)
+            {
+                if (isProcessed(ii) || monitoredItems[ii] == null)
+                {
+                    continue;
+                }
+
+                IAsyncNodeManager owner = monitoredItems[ii].NodeManager;
+                if (owner is null)
+                {
+                    continue;
+                }
+
+                owners ??= [];
+                ownerIndex ??= new Dictionary<object, int>(RefEqualityComparer.Default);
+                if (!ownerIndex.TryGetValue(owner, out int group))
+                {
+                    owners.Add((owner, []));
+                    group = owners.Count - 1;
+                    ownerIndex[owner] = group;
+                }
+
+                owners[group].Indices.Add(ii);
+            }
+
+            return owners;
+        }
+
+        /// <summary>
+        /// Dispatches an ownership-sensitive data monitored item operation to each item's
+        /// recorded owning NodeManager rather than to the visible routing-table managers.
+        /// Each owner is offered only the items it owns (all other indices are pre-marked
+        /// processed) so a same-namespace replacement generation can never claim monitored
+        /// items still owned by a shadow-retired generation. Owners that are no longer
+        /// registered in the routing table are shadow-retired generations; when
+        /// <paramref name="notifyRetiredGenerationDrain"/> is set the registered drain
+        /// observer is notified afterwards so retired generations can be torn down once
+        /// their monitored items drain.
+        /// </summary>
+        private async ValueTask DispatchDataMonitoredItemsToOwningNodeManagersAsync(
+            IList<IMonitoredItem> monitoredItems,
+            List<bool> processedItems,
+            Func<IAsyncNodeManager, IList<bool>, ValueTask> dispatch,
+            bool notifyRetiredGenerationDrain,
+            CancellationToken cancellationToken)
+        {
+            List<(IAsyncNodeManager Owner, List<int> Indices)>? owners =
+                GroupDataMonitoredItemsByOwner(
+                    monitoredItems,
+                    index => processedItems[index]);
+            if (owners is null)
+            {
+                return;
+            }
+
+            bool retiredGenerationDrained = false;
+            foreach ((IAsyncNodeManager owner, List<int> indices) in owners)
+            {
+                // Present only this owner's items as unprocessed.
+                bool[] ownedItems = new bool[monitoredItems.Count];
+                for (int ii = 0; ii < ownedItems.Length; ii++)
+                {
+                    ownedItems[ii] = true;
+                }
+                foreach (int ii in indices)
+                {
+                    ownedItems[ii] = false;
+                }
+
+                await dispatch(owner, ownedItems).ConfigureAwait(false);
+
+                // Merge the owner's processed marks back into the shared list.
+                foreach (int ii in indices)
+                {
+                    if (ownedItems[ii])
+                    {
+                        processedItems[ii] = true;
+                    }
+                }
+
+                if (notifyRetiredGenerationDrain && !m_nodeManagers.Contains(owner))
+                {
+                    retiredGenerationDrained = true;
+                }
+            }
+
+            if (retiredGenerationDrained)
+            {
+                m_owner.NotifyRetiredGenerationDrainObserver();
+            }
+        }
+
+        /// <summary>
+        /// Dispatches Modify to each data monitored item's recorded owning NodeManager.
+        /// Modify tracks per-item completion through <see cref="MonitoredItemModifyRequest.Processed"/>
+        /// rather than a processed-flag list, so each owner is isolated by temporarily
+        /// marking every item it does not own as processed for the duration of its call.
+        /// </summary>
+        private async ValueTask DispatchModifyToOwningNodeManagersAsync(
+            OperationContext context,
+            TimestampsToReturn timestampsToReturn,
+            IList<IMonitoredItem> monitoredItems,
+            ArrayOf<MonitoredItemModifyRequest> itemsToModify,
+            IList<ServiceResult> errors,
+            IList<MonitoringFilterResult> filterResults,
+            CancellationToken cancellationToken)
+        {
+            List<(IAsyncNodeManager Owner, List<int> Indices)>? owners =
+                GroupDataMonitoredItemsByOwner(
+                    monitoredItems,
+                    index => itemsToModify[index].Processed);
+            if (owners is null)
+            {
+                return;
+            }
+
+            foreach ((IAsyncNodeManager owner, List<int> indices) in owners)
+            {
+                bool[] ownedItems = new bool[monitoredItems.Count];
+                foreach (int ii in indices)
+                {
+                    ownedItems[ii] = true;
+                }
+
+                // Temporarily mark every item this owner does not own as processed so it
+                // only touches its own items, then restore them for the next owner.
+                var masked = new List<int>();
+                for (int ii = 0; ii < monitoredItems.Count; ii++)
+                {
+                    if (!ownedItems[ii] && !itemsToModify[ii].Processed)
+                    {
+                        itemsToModify[ii].Processed = true;
+                        masked.Add(ii);
+                    }
+                }
+
+                try
+                {
+                    await owner.ModifyMonitoredItemsAsync(
+                            context,
+                            timestampsToReturn,
+                            monitoredItems,
+                            itemsToModify,
+                            errors,
+                            filterResults,
+                            cancellationToken)
+                        .ConfigureAwait(false);
+                }
+                finally
+                {
+                    foreach (int ii in masked)
+                    {
+                        itemsToModify[ii].Processed = false;
+                    }
+                }
+            }
+        }
+
+        internal async ValueTask<ServiceResult?> ValidateMonitoredItemCreateRequestAsync(
+            OperationContext operationContext,
+            MonitoredItemCreateRequest item,
+            CancellationToken cancellationToken = default)
+        {
+            // check for null structure.
+            if (item == null)
+            {
+                return new ServiceResult(StatusCodes.BadStructureMissing);
+            }
+
+            // validate read value id component. Validate also access rights and permissions
+            ServiceResult? error = await ValidateReadRequestAsync(
+                    operationContext,
+                    item.ItemToMonitor,
+                    null,
+                    true,
+                    cancellationToken)
+                .ConfigureAwait(false);
+
+            if (ServiceResult.IsBad(error))
+            {
+                return error;
+            }
+
+            // check for valid monitoring mode.
+            if ((int)item.MonitoringMode is < 0 or > ((int)MonitoringMode.Reporting))
+            {
+                return new ServiceResult(StatusCodes.BadMonitoringModeInvalid);
+            }
+
+            // check for null structure.
+            MonitoringParameters attributes = item.RequestedParameters;
+
+            error = MasterNodeManager.ValidateMonitoringAttributes(attributes);
+
+            if (ServiceResult.IsBad(error))
+            {
+                return error;
+            }
+
+            // check that no filter is specified for non-value attributes.
+            if (item.ItemToMonitor.AttributeId is not Attributes.Value and not Attributes
+                .EventNotifier)
+            {
+                if (!attributes.Filter.IsNull)
+                {
+                    return new ServiceResult(StatusCodes.BadFilterNotAllowed);
+                }
+            }
+            else
+            {
+                error = MasterNodeManager.ValidateMonitoringFilter(attributes.Filter);
+
+                if (ServiceResult.IsBad(error))
+                {
+                    return error;
+                }
+            }
+
+            // passed basic validation.
+            return null;
+        }
+
+        internal async ValueTask<ServiceResult> ValidateCallRequestItemAsync(
+            OperationContext operationContext,
+            CallMethodRequest callMethodRequest,
+            Dictionary<NodeId, Variant[]>? uniqueNodesReadAttributes = null,
+            bool permissionsOnly = false,
+            CancellationToken cancellationToken = default)
+        {
+            // check for null structure.
+            if (callMethodRequest == null)
+            {
+                return StatusCodes.BadStructureMissing;
+            }
+
+            // check object id.
+            if (callMethodRequest.ObjectId.IsNull)
+            {
+                return StatusCodes.BadNodeIdInvalid;
+            }
+
+            // check method id.
+            if (callMethodRequest.MethodId.IsNull)
+            {
+                return StatusCodes.BadMethodInvalid;
+            }
+
+            (_, IAsyncNodeManager? nodeManager) = await m_owner.GetManagerHandleAsync(callMethodRequest.ObjectId, cancellationToken)
+                    .ConfigureAwait(false);
+
+            if (nodeManager == null)
+            {
+                return StatusCodes.BadNodeIdUnknown;
+            }
+
+            // Method resolution fallback is provided by the IAsyncNodeManager adapter path.
+            MethodState method = await nodeManager.FindMethodStateAsync(
+                operationContext,
+                callMethodRequest,
+                cancellationToken).ConfigureAwait(false);
+
+            if (method != null)
+            {
+                // check access rights and role permissions
+                return await ValidatePermissionsAsync(
+                        operationContext,
+                        method.NodeId,
+                        PermissionType.Call,
+                        uniqueNodesReadAttributes,
+                        permissionsOnly,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+            }
+
+            return StatusCodes.Good;
+        }
+
+        internal async ValueTask<ServiceResult?> ValidateReadRequestAsync(
+            OperationContext operationContext,
+            ReadValueId readValueId,
+            Dictionary<NodeId, Variant[]>? uniqueNodesReadAttributes = null,
+            bool permissionsOnly = false,
+            CancellationToken cancellationToken = default)
+        {
+            ServiceResult? serviceResult = ReadValueId.Validate(readValueId);
+
+            if (ServiceResult.IsGood(serviceResult))
+            {
+                //any attribute other than Value or RolePermissions
+                PermissionType requestedPermission = PermissionType.Browse;
+                if (readValueId.AttributeId == Attributes.RolePermissions)
+                {
+                    requestedPermission = PermissionType.ReadRolePermissions;
+                }
+                else if (readValueId.AttributeId == Attributes.Value)
+                {
+                    requestedPermission = PermissionType.Read;
+                }
+
+                // check access rights and role permissions
+                serviceResult = await ValidatePermissionsAsync(
+                        operationContext,
+                        readValueId.NodeId,
+                        requestedPermission,
+                        uniqueNodesReadAttributes,
+                        permissionsOnly,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+            }
+            return serviceResult;
+        }
+
+        internal async ValueTask<ServiceResult?> ValidateWriteRequestAsync(
+            OperationContext operationContext,
+            WriteValue writeValue,
+            Dictionary<NodeId, Variant[]>? uniqueNodesServiceAttributes = null,
+            bool permissionsOnly = false,
+            CancellationToken cancellationToken = default)
+        {
+            ServiceResult? serviceResult = WriteValue.Validate(writeValue);
+
+            if (ServiceResult.IsGood(serviceResult))
+            {
+                PermissionType requestedPermission = PermissionType.WriteAttribute; //any attribute other than Value, RolePermissions or Historizing
+                if (writeValue.AttributeId == Attributes.RolePermissions)
+                {
+                    requestedPermission = PermissionType.WriteRolePermissions;
+                }
+                else if (writeValue.AttributeId == Attributes.Historizing)
+                {
+                    requestedPermission = PermissionType.WriteHistorizing;
+                }
+                else if (writeValue.AttributeId == Attributes.Value)
+                {
+                    requestedPermission = PermissionType.Write;
+                }
+
+                // check access rights and permissions
+                serviceResult = await ValidatePermissionsAsync(
+                        operationContext,
+                        writeValue.NodeId,
+                        requestedPermission,
+                        uniqueNodesServiceAttributes,
+                        permissionsOnly,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+            }
+            return serviceResult;
+        }
+
+        internal async ValueTask<ServiceResult?> ValidateHistoryReadRequestAsync(
+            OperationContext operationContext,
+            HistoryReadValueId historyReadValueId,
+            CancellationToken cancellationToken = default)
+        {
+            ServiceResult? serviceResult = HistoryReadValueId.Validate(historyReadValueId);
+
+            if (ServiceResult.IsGood(serviceResult))
+            {
+                // check access rights and permissions
+                serviceResult = await ValidatePermissionsAsync(
+                        operationContext,
+                        historyReadValueId.NodeId,
+                        PermissionType.ReadHistory,
+                        null,
+                        true,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+            }
+            return serviceResult;
+        }
+
+        internal async ValueTask<ServiceResult?> ValidateHistoryUpdateRequestAsync(
+            OperationContext operationContext,
+            HistoryUpdateDetails historyUpdateDetails,
+            CancellationToken cancellationToken = default)
+        {
+            ServiceResult? serviceResult = HistoryUpdateDetails.Validate(historyUpdateDetails);
+
+            if (ServiceResult.IsGood(serviceResult))
+            {
+                // check access rights and permissions
+                PermissionType requiredPermission = MasterNodeManager.DetermineHistoryAccessPermission(
+                    historyUpdateDetails);
+                serviceResult = await ValidatePermissionsAsync(
+                    operationContext,
+                    historyUpdateDetails.NodeId,
+                    requiredPermission,
+                    null,
+                    true,
+                    cancellationToken).ConfigureAwait(false);
+            }
+
+            return serviceResult;
+        }
+
+        internal async ValueTask<ServiceResult> ValidatePermissionsAsync(
+            OperationContext context,
+            NodeId nodeId,
+            PermissionType requestedPermision,
+            Dictionary<NodeId, Variant[]>? uniqueNodesServiceAttributes = null,
+            bool permissionsOnly = false,
+            CancellationToken cancellationToken = default)
+        {
+            if (context.Session != null)
+            {
+                (object? nodeHandle, IAsyncNodeManager? nodeManager) = await m_owner.GetManagerHandleAsync(nodeId, cancellationToken)
+                    .ConfigureAwait(false);
+
+                return await ValidatePermissionsAsync(
+                        context,
+                        nodeManager,
+                        nodeHandle,
+                        requestedPermision,
+                        uniqueNodesServiceAttributes,
+                        permissionsOnly,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+            }
+            return StatusCodes.Good;
+        }
+
+        internal async ValueTask<ServiceResult> ValidatePermissionsAsync(
+            OperationContext context,
+            IAsyncNodeManager? nodeManager,
+            object? nodeHandle,
+            PermissionType requestedPermision,
+            Dictionary<NodeId, Variant[]>? uniqueNodesServiceAttributes = null,
+            bool permissionsOnly = false,
+            CancellationToken cancellationToken = default)
+        {
+            (ServiceResult result, _) = await ValidatePermissionsAndGetMetadataAsync(
+                context,
+                nodeManager,
+                nodeHandle,
+                requestedPermision,
+                uniqueNodesServiceAttributes,
+                permissionsOnly,
+                metadataRequired: false,
+                cancellationToken).ConfigureAwait(false);
+            return result;
+        }
+
+        internal async ValueTask<(ServiceResult result, NodeMetadata? metadata)>
+            ValidatePermissionsAndGetMetadataAsync(
+                OperationContext context,
+                IAsyncNodeManager? nodeManager,
+                object? nodeHandle,
+                PermissionType requestedPermission,
+                Dictionary<NodeId, Variant[]>? uniqueNodesServiceAttributes,
+                bool permissionsOnly,
+                bool metadataRequired,
+                CancellationToken cancellationToken)
+        {
+            if (nodeManager == null ||
+                nodeHandle == null ||
+                (context.Session == null && !metadataRequired))
+            {
+                return (StatusCodes.Good, null);
+            }
+
+            // First attempt to retrieve just the Permission metadata with or without cache optimization
+            // If it happens that nodemanager does not fully implement GetPermissionMetadata,
+            // fallback to GetNodeMetadataAsync
+            NodeMetadata? nodeMetadata = await nodeManager.GetPermissionMetadataAsync(
+                    context,
+                    nodeHandle,
+                    BrowseResultMask.NodeClass,
+                    uniqueNodesServiceAttributes!,
+                    permissionsOnly,
+                    cancellationToken)
+                .ConfigureAwait(false);
+
+            // If GetPermissionMetadataAsync returns null, or a caller needs a NodeClass
+            // that the optimized metadata path did not populate, read the full metadata.
+            if (nodeMetadata == null ||
+                (metadataRequired && nodeMetadata.NodeClass == NodeClass.Unspecified))
+            {
+                NodeMetadata? fullMetadata = await nodeManager.GetNodeMetadataAsync(
+                        context,
+                        nodeHandle,
+                        BrowseResultMask.NodeClass,
+                        cancellationToken)
+                    .ConfigureAwait(false);
+                nodeMetadata = fullMetadata ?? nodeMetadata;
+            }
+
+            if (nodeMetadata == null || context.Session == null)
+            {
+                return (StatusCodes.Good, nodeMetadata);
+            }
+
+            return (
+                ValidatePermissionMetadata(context, nodeMetadata, requestedPermission),
+                nodeMetadata);
+        }
+
+        internal ServiceResult ValidatePermissionMetadata(
+            OperationContext context,
+            NodeMetadata nodeMetadata,
+            PermissionType requestedPermission)
+        {
+            ServiceResult result = MasterNodeManager.ValidateRolePermissions(
+                context,
+                nodeMetadata,
+                requestedPermission,
+                m_logger);
+            return ServiceResult.IsGood(result)
+                ? MasterNodeManager.ValidateAccessRestrictions(context, nodeMetadata)
+                : result;
+        }
+
+        private readonly MonitoredItemIdFactory m_monitoredItemIdFactory = new();
+        private readonly MasterNodeManager m_owner;
+        private readonly NodeManagerRoutingTable m_nodeManagers;
+        private readonly ILogger m_logger;
+    }
+}
